[iOS] Make ErrorPageHelper compatible with framework
This CL makes sure that the ErrorPageHelper and the new error workflow
plays nicely with the framework bundle.
It adds the CRW prefix to the class and changes the bundle in which the
error page HTML files are search to the bundle containing the helper
class.

Bug: 1152364
Change-Id: Ibb6e07d5143bb7527d7f9ad74a59b20599e38899
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2563667
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Eugene But <eugenebut@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#832363}
GitOrigin-RevId: 30154cd21a359d11b10d998f788d85cddca508c5