Plumb base::Optional<SessionID> to CreateHistoricalTab() and AddRestoredTab
ClosedTabCache is a desktop feature to instantly restore the
closed tabs. To able to do that, each tab should be identified by
a unique id which in our case is SessionID.

To accommodate this, we need to send the SessionID whenever user
tries to restores the tab using AddRestoredTab.

This CL plumbs through CreateRestoredTab which restores the WebContents
from the cache in the event of a cache hit. Now passing SessionID as
an optional argument through the function calls for AddRestoredTab.

BUG=1100946

Change-Id: I05366c8ef3ced9b901aa5c124d85b025a7fd463f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2564812
Commit-Queue: Sreeja Kamishetty <sreejakshetty@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/master@{#832718}
GitOrigin-RevId: 673efca650fdece808b89c6e75392e1ecb4d4fd9