[iOS] Fix WebView tests with new error pages
This CL fixes the WebView tests when the new error pages are enabled.
The test needs to wait for the error page javascript to be injected in
the page before finishing, otherwise the webstate is deallocated
before the end.
To do that, the test is waiting for the error string to be inside the
web view.

Bug: 1152364
Change-Id: I3f958c2fc86b881a1a3f5f81fbe97b68a6e40cd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2565534
Reviewed-by: Eugene But <eugenebut@chromium.org>
Reviewed-by: John Wu <jzw@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#832837}
GitOrigin-RevId: e8bcc73aa3eb0cf805434cdad98747b428d87792