[iOS] Fix over-recording of WindowActivityExternalOrigin
WindowActivityExternalOrigin is being recorded incorrectly in the
IOS.MultiWindow.OpenInNewWindow histogram in two cases:

- When the app launches and creates a window using the Scene APIs for
  the first time, including on iPhones.

- When the user drags a tab out of the tab grid to create a new window.

The fix for the first case involves adding more logic to skip logging
window creation: not if multiple scenes aren't supported, and not if
there's zero or one scenes (as there would be during the initial launch).

The fix for the second case is just to add a window origin value to the
user activity generated for a drag action.

Additionally, this CL adds missing enum values to the XML so it now
matches the WindowActivityOrigin enum.

Bug: 1153130
Change-Id: I1b49aea19bad67e80b5b9542ef72c226d9a46412
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2563301
Reviewed-by: David Jean <djean@chromium.org>
Commit-Queue: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#832899}
GitOrigin-RevId: 1ba4ee896dfdd6349fe7cb1706b465a98b8fc23e