[iOS] Creates NewTabPageViewController
Creates a view controller for the NTP, the NewTabPageViewController.
This VC is managed by the NewTabPageCoordinator and contains all the
content that is presented on the regular NTP.

First step of the refactoring effort; more content will be added to it
in later CLs.

Bug: 1114792
Change-Id: I38ce3fc3b386f9d93deed4ca28c9ec879b47a781
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2555662
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/master@{#833889}
GitOrigin-RevId: c4e0600191bf49a6df0cc351d9a235154de4983e