Introduce PrimaryAccountMutator::RevokeConsent
The existing PrimaryAccountMutator::ClearPrimaryAccount with options
to keep or remove all accounts was very confusing and it lead to bug
http://crbug.com/1155042

This CL fixes the public API with 2 clear methods:
* RevokeSyncConsent: replaces the version to clear the primary account
when with kDefault action and leaves the primary account in
ConsentLevel::kNotRequired on ChromeOS and DICE.
* ClearPrimaryAccount: clears the primary account (revokes all consents) and
also removes all accounts.

A series of follow-up CLs will clean the internals of the
PrimaryAccountManager as they become obsolete.

Some additional considerations:
* iOS and Android do *not* use RevokeSyncConsent(), except from the
unexpected call from ProfileSyncService. We could ifdef out
RevokeSyncConsent() on iOS and Android, which will force callers like
the ProfileSyncService to simply force clear the primary account
when this happens.
* ChromeOS ASH cannot clear the primary account, so the method
ClearPrimaryAccount() does not exist on this platform.

Bug: 1155505

Change-Id: I9e8b0764599d38d53063764695d91d29de05d154
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2573082
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#834928}
GitOrigin-RevId: 3a5fd2389a1459ad2e60ca4e8ab5dfbb78f6252e