[ios] Add histogram to investigate loss off UKM.LogSize.OnSuccess records
There is a significant discrepancy between number of UKM records
uploaded to google servers and numbder of UKM.LogSize.Success records.
This either means that UMA data was not written to disk or that app was
terminated after server received the data, but before UMA was recorded.
Data loss caused by failure to write on disk is highly plausible if the
app is running in the background.

This CL adds another metric, which counts the number of times
UKM.LogSize.Success is emitted. The counter is stored in NSUserDefaults
(hence have higher chance to persist). The counter is recorded as UMA
when the app starts or foregrounded, hence have higher changes to get
written to disk.

Bug: 1154678
Change-Id: Ifa71355e02f568713f9d87f45692ae70fd496b72
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2576077
Auto-Submit: Eugene But <eugenebut@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#835191}
GitOrigin-RevId: ea6f746bf472b2031f2e4ff8cd5378238272c731