[iOS Enterprise] BrowserSignin: hide promos
Makes the sign-in promo view mediator aware of the kSigninAllowed pref
and makes it skip the sign-in promos in the bookmarks and the settings
if sign-in is disabled by policy.

Also makes a similar change in the sign-in utils so the sign-in promo
after a version upgrade is skipped if sign-in is disabled by policy.

Change-Id: Ia84e43f9db317f9454d662ea998758be66ffa801
Bug: 1155745
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2562778
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#835793}
GitOrigin-RevId: 902a4b1e3cd4da60a7ce893ff2ee4a212f285bf7