[iOS] Fix thumb strip thumbnails overlapping status bar
The major issue here was that the height calculation in
horizontal_layout.mm wasn't taking the collection view's contentInsets
into account when calculating the sectionInset. This meant that the
resulting layout was too large. Specifically, the vertical sectionInsets
+ contentInset + item height ended up larger than the collection view's
height, causing the layout to be incorrect.

Fixed: 1155601
Change-Id: Ie9594fec2762239963bc90c41bf07f8caa42bf83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2587038
Reviewed-by: edchin <edchin@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#836196}
GitOrigin-RevId: c1be048d177189bc61533dae278e61ee12d8a691