[iOS] Custom UIScrollViewDelegate for feed
Upstream implementation for adding a custom UIScrollViewDelegate to
the feed. Uses NewTabPageCoordinator as UIScrollViewDelegate for the
feed collection view. This allows for us to monitor scroll actions in
the NTP when the feed is the primary scroll view.

Downstream: crrev.com/c/3466483
google3: cl/347390497

Bug: 1114792
Change-Id: I0b9fc60f6de1856270bdda7cf71ce4dcc40e3bb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2588639
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/master@{#836869}
GitOrigin-RevId: 2211692785f58b5b5f8c5bc86491c7567c9dac44