[iOS] Disable the site info tools menu when the web content is blocked
crrev.com/c/2136232 added the ability to block the web content area
when an HTTP auth dialog from a different host is displayed. However,
in this case, the tools menu site info item is still enabled and shows
the incorrect url. This disables that info when the content area is
blocked.

Bug: 1157800
Change-Id: I0004c0ce608303e41e970f4389481ad31bcd050d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2592998
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#837587}
GitOrigin-RevId: 8f03df37d50bcef05319223815344154659e0ad3