[iOS][QRScanner] Fix text animation
Worked on this doing pair programming with Sébastien today, and after
debugging we found out that the change in the width of the text is what
causes the animation to appear like it "jumps". Having the text width
to be constant fixes that, and looks good on normal dynamic size fonts.
However it does needs scrolling on the bigger a11y fonts. Would like to
get your input in the way i'm fixing the bug, as this will decrease the
horizontal space for text on iPads.

Another issue we found was that on compact height the text is cropped
into <1 line, and looks bad. Hiding it fixes that.

Fixed: 1005249
Change-Id: I0651ec85ddbb5bbbf30f50518bbb9c6be7f9392a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2595857
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838126}
GitOrigin-RevId: 68b0d770acbe32e44d909a960c4a4bc30ca09160