[iOS Enterprise] BrowserSignin: Disable settings entry point
Adds logic to the settings view controller to replace the sign-in item
with an info button item that grays out the sign-in option. The info
button's popover menu explains that sign-in is disable by the managed
browser's organization, and provides a link to the chrome://management
page for more information.

As part of this change, the TableViewInfoButtonItem class was modified
to support changing the color of the cell's main text and letting
callers provide a pre-instantiated UIImage. Passing an image rather
than an image name is necessary in the case of sign-in, because the
image must be created by the SigninResourcesProvider.

In addition, the EnterpriseInfoPopoverViewController was modified to
enable one of the disabled constructors, which lets callers provide
the primary text message.

Screenshots of the new UI when sign-in is disabled:
https://screenshot.googleplex.com/38X86W24GXm2Ej5
https://screenshot.googleplex.com/9yee95xBSJ7u3D6

UI mocks:
go/bling-enterprise-q42020 (slide 6)

Bug: 1155745
Change-Id: I099c67b93023fb3f78181bdfeae13dcb73a0e59e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2591388
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838195}
GitOrigin-RevId: 77847d0ce90a8eb04893ac30bae52158a99105fd