[ios] Fix session restoration issues when using Scene API
When Scene API is used (which requires a device recent enough running
iOS 13.0 or later), UIApplicationDidEnterBackgroundNotification is not
sent to the application if it is terminated by swiping it away while
it is running in the foreground.

TabModel listened to UIApplicationDidEnterBackgroundNotification to
save the session when the app is backgrounded or terminated. With the
Scene API, this save no longer occurred when the app was terminated.

The session was still saved in two other circumstances:
1. when the active tab navigation was committed (immediately),
2. when the active tab was changed (delayed by ~ two seconds).

Those two other time where the session could be saved made this bug
only exhibit in really specific circumstances (i.e. closing a tab
that was not active, then terminating the app while foregrounded).

To fix this, schedule a delayed saving of the session every when
the WebStateList changes (i.e. tab is detached, added, moved or
replaced) and save immediately when the TabModel is disconnected
(which happens when the Scene is closed or the app it terminated).

Bug: 1146402, 1152751
Change-Id: I64b17a4b0a9417c629daad565703a399a63c229a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2599858
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838841}
GitOrigin-RevId: 92117db0be1b7266bde1b0cbd67c7d35b968dffd