[iOS] Ensure that Sync tests are more flexible on startup.
Currently, Chrome strictly enforces that the fake Sync server is running
to call |ClearSyncServerData|. This may be causing intermittent issues
in EarlGrey tests if service startup is in an inconsistent state.

This patch makes the test suites more flexible by allowing the
inconsistent state, and changing a failure for a no-op case.

Bug: 1153087
Change-Id: I4f20dc0c282d1b09ad54d43ab32dd10f54a2d42c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2600741
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#838958}
GitOrigin-RevId: 9a6e44591e27714ec29f0d78bc28b8b34a25f6f9