[ios] Ignores newSuggestionsInCategory if Discover is enabled.
The crash from 1145109 seems to happens due to an update from
newSuggestionsInCategory triggering a layoutSubviews pass on
Discover.

Since ContentSuggestionsServiceObserver is used for the old content
suggestions system (and not Discover's) this callback will now be
ignored if the Discover feed is being used.

Bug: 1145109
Change-Id: Iec40e155028f6406e702147371d208dc69471efc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2605284
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#839456}
GitOrigin-RevId: f265708066f271cc521e9ed4bd18b9fe3050ec2a