Default typed omnibox navigations to HTTPS: Initial implementation
Presently, when a user types a domain name in the omnibox such as
"example.com", Chrome navigations to the HTTP version of the site
(http://example.com). However, the web is increasingly moving towards
HTTPS, and we now want to optimize omnibox navigations and first-load
performance for HTTPS, rather than HTTP.

This CL implements an initial version of defaulting typed omnibox
navigations to HTTPS. In particular, the CL:
- Modifies omnibox and autocomplete code to use HTTPS as the default
  scheme for navigations. We call these "upgraded HTTPS navigations".
- Observes upgraded HTTPS navigations until completion and falls back
  to the HTTP version of the URL if the HTTPS load fails. It does this
  by introducing a new navigation throttle called
  TypedNavigationUpgradeThrottle.
 - Ignores SSL errrs in SSLErrorNavigationThrottle if the navigation was
  an HTTPS upgrade in order to prevent displaying spurious SSL
  interstitials.

Most of the files touched by the CL have plumbing changes: The
navigation throttle needs to know whether the navigation was upgraded,
so this information is provided by the omnibox via
is_using_https_as_default_scheme bit and stored inside the UIData of the
navigation handle.

This is a minimal implementation and is not ready for general usage.
Future CLs are going to observe upgraded HTTPS navigations for several
seconds instead and cancel the load when necessary, instead of
indefinitely waiting for HTTPS loads to succeed. This CL also lacks many
quality of life improvements such as remembering which URLs fell back to
HTTP. These will also be added in future CLs.

Bug: 1141691
Change-Id: Ie78a34bfa51fc995e9834d491ea43b7cac17598c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2568448
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Clark DuVall <cduvall@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840253}
GitOrigin-RevId: 06028d794a060329e3d356d55f8539035b4e57eb