[iOS Enterprise] BrowserSignin: disable sync entry point
Adds logic to the Sync & Google Services view mediator to disable the
sign-in button when browser sign-in is disabled by policy. The info
button's popover menu explains that sign-in is disable by the managed
browser's organization, and provides a link to the chrome://management
page for more information.

No special handling is needed for the mobile identity consistency
experiment because when it is enabled, the sign-in entry point in
Google Services is removed from the UI.

Screenshots of the new UI when sign-in is disabled:
https://screenshot.googleplex.com/38iQWPYitD7hKCo
https://screenshot.googleplex.com/4ozvEWnJLqdnzsY

Bug: 1155745
Change-Id: I9348e6ae724d40ab8b38c39f45726bac0c8f5fa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2594011
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#840770}
GitOrigin-RevId: 4675b70d6638eb3bcc575802fd03a6f54ae5eec5