[iOS][ThumbStrip] Set BVC position on new tabs if using smooth scrolling
Now, instead of having the BVC manually set the contentInset, it
instead alerts the fullscreen controller that the thumb strip has
appeared. The fullscreen controller can then use this to correctly set
the insets itself, even when new tabs are added.

Also, the BVC now keeps track of whether the thumb strip has caused
the view to be translated, so it can re-apply this translation when
adding a new web state.

Fixed: 1155160
Change-Id: I827d9edf058667a535275367605518512ebca7e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2599866
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#841461}
GitOrigin-RevId: 3ab83c88acdbd095c68efcd77d4a79b9e793cf35