[iOS] Add UIContextMenu support for context menu
This CL adds a new class that could handle using UIContextMenu to
display context menu.
The new class is reusing the existing JavaScript from the current
context menu to find the element on which the long press was done.

For now the class isn't connected and isn't displaying anything.

Bug: 1140387
Change-Id: I33c2f90445af29712b8680da3f69cdda21dd0367
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2578958
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#841532}
GitOrigin-RevId: 9ca289f6a2e5cd324f91945684d2b8a1a2e29732