Reland "Abstract signin dep from RealTimeUrlLookupService"
This is an unchanged reland of r#842942. That CL was reverted due to
suspicions of it causing SubresourceRedirectLoggedInSitesBrowserTest.TestCancelBeforeImageLoadForLoggedInSite
to flake. However, it turns out that that test was already flaky, as can
be seen on the flake portal. For example, this run shows it failing on a
CQ run of r#842899:
https://ci.chromium.org/ui/p/chromium/builders/ci/Linux%20Tester%20(Ozone%20Wayland)/4088/overview

Original CL description:

This CL continues with the abstraction of dependencies on signin and
sync from //components/safe_browsing's access token fetching flow in
order to enable reuse by WebLayer. The concrete step taken here is to
abstract the dependencies on
//components/safe_browsing/core/browser/sync from
RealTimeUrlLookupService, passing these dependencies in via the
constructor. There are no behavioral changes in this CL:
- //chrome and //ios/chrome's factories glue the relevant parameters
  from //components/safe_browsing/core/browser/sync
- //weblayer's factory passes in values for the parameters such that
  access token fetching for URL lookups continues to be disabled in
  WebLayer

Followup work will bring up support for access token fetching in
WebLayer.

Bug: 1080748
TBR: kolos@chromium.org
Change-Id: I373f970373de7c5d14568a165365fe5a445aef08
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2627489
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Maxim Kolosovskiy  <kolos@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843052}
GitOrigin-RevId: 57009ca2357570399628f12a59a4ad00453493c9