[ios] Refactor creation of Browser in BrowserViewWrangler
The code responsible for the creation of the regular and incognito
Browsers was duplicated. Introduce an helper method that creates a
Browser associated with a BrowserState and sets it up, then change
BrowserViewWrangler to use this method for creating both Browsers.

Bug: 1165798
Change-Id: I7a02f7fb8bc3e724ee9064a29956ecd316074e69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2627318
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: sebsg <sebsg@chromium.org>
Reviewed-by: sebsg <sebsg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843239}
GitOrigin-RevId: 2b72695a1ee7380794cf5e160679562a32205b44