[Safe Browsing] Abstract sync dep from RealTimeURLLookupServiceBase
This CL completes the abstraction of signin & sync deps from
//components/safe_browsing/core/realtime by abstracting the usage of
SyncService to determine whether history sync is enabled. As with the
other abstracted dependencies, this information is now obtained via a
callback that is passed in from the embedder. With this CL, WebLayer
can now use this component without needing to reference signin or sync
at all.

There is no behavioral change in this CL.

Bug: 1080748
Change-Id: I320e7f225fb65258862768e411dd85c51dfac9be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2624211
Commit-Queue: Colin Blundell <blundell@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#843476}
GitOrigin-RevId: 86becbf073d80ec5337a23f9f6734251f2f2af72