[iOS][MICe] Fix UI crash when signing in to account with Sync password.
This patch fixes a crash that occurs when a user goes through the
advanced Sync settings sign-in flow with an account that has a Sync
encryption password.

Starting from this patch Sync error cards are only displayed after the
sign-in flow has been completed.

Bug: 1125631
Change-Id: I877e59e25265e78a4220fbb66b7a3855757dceb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2637575
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844771}
GitOrigin-RevId: df71978f06d1569de8973accb663cbf66522adb2