Use DistilledURL when loading distilled data
When using loadData, the html has access to the cookies of the URL.
It is necessary to pass the URL of the distilled page to avoid cross
domain leak.

This CL breaks the workflow of user following a new redirection on
reloading the page.

Bug: 1167901
Change-Id: I6cb01a6580ed511de19f55b35acf91f7f3c7493d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2637584
Auto-Submit: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#844982}
GitOrigin-RevId: c0015aa91f3a587c6a43e73bed963f75597b3a48