[chrome://policy] Implement handling for "Superseded" policy conflicts
When a policy conflicts with another and their values are the same, the
policy page now displays a message under the status "Info", as well as
labeling the conflict as "Superseding". The wording of the conflict
status messages is also adjusted to increase clarity. Screenshot SHAs
depicting the new labels and messages are included in this CL.

In addition, support for various conflict types is implemented by adding
a new EntryConflict class. The conflict type is determined based on the
two conflicting policies' values. If the values are equal, the conflict
is treated as superseding. Otherwise, the conflict is overriding.

Policy table before update: http://screen/7RMnvFaR3S3xcC6
Policy table after update: http://screen/5JE3yrUsn5JKsVT

Bug: 1104043
Change-Id: I36dea4b14f8068854d8d7713892187d930956e23
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2625896
Commit-Queue: Igor Ruvinov <igorruvinov@chromium.org>
Reviewed-by: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Yann Dago <ydago@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845802}
GitOrigin-RevId: 6b01e4a73b805c61a4e63233370f7ae2d97167e2