[ios] Use base::FilePath for SessionServiceIOS/CrashRestoreHelper APIs
Requiring the BrowserState's path to be passed as a NSString* forces
all callers to convert the base::FilePath to NSString* making the API
harder to use (as the conversion is non trivial).

Instead change the API of both classes to use base::FilePath directly
and to do the conversion internally where it can be centralized in a
few location makes the API easier to use.

Bug: 1165798
Change-Id: Ida8294210f83875fcbeffa88a3e7f5ae78a2316f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2635668
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#846070}
GitOrigin-RevId: 559742ff90a51253bcb104cb0c2cd6ab323fab6c