[iOS] Fix Fakebox width in screenshots
This CL ensures that the fakebox width is correct when the screenshot of
the NTP is taken. It was wrong as the width was first computed while the
collection view had a zero width. The width is now re-computed when the
NTP is added to its parent ViewController.

Fixed: 1165956
Change-Id: I63cf57584fe185280956779791973495092dfbb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2640440
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#846291}
GitOrigin-RevId: 61938138fdcdc866d90b70ed0291075f3af08cf3