[iOS] Fix issue where refactored NTP scrolling jumps down
Prevents the NTPVC from being set as a child of the BVC multiple times
by removing any parent VC in willMoveToParentViewController.

This would cause the scroll position to not be in sync with what it
looked like, and would jump when scrolling any amount.

Bug: 1169954
Change-Id: Ic811df0b1fb61d0ba2a48107a87e709f0d921be5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2643062
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847500}
GitOrigin-RevId: 3431db30aca287c27249082d7a2bd24ff203201a