Invalidations: Get active account to use from IdentityManager, not Sync
Before this CL, components/invalidation relied on SetActiveAccountId()
calls from Sync to determine which account to use. This could cause
issues in one particular situation: If LocalSync (aka roaming profile
support) is enabled, Sync doesn't use any account, and so wouldn't call
SetActiveAccountId(). In this situation, the invalidations component
would remain inactive, and other clients of it might not work correctly.

The SetActiveAccountId() API was also a bad abstraction:
- Out of the two IdentityProvider implementations, only one (Profile..)
  actually supports it. The other one (Device..) ignores the passed-in
  account (apart from logging if it doesn't match the expected value).
- components/invalidation has other clients besides Sync; it's weird
  that Sync specifically should be the one to determine the account.

The reason for this setup is historic: This was built before the signin
code itself knew the concept of an "unconsented primary account" (which
is the account that Sync uses).
The current IdentityManager API supports this just fine though, and
ProfileIdentityProvider even already observes the IdentityManager.

So this CL removes the SetActiveAccountId() API, and instead gets the
account to use from IdentityManager. As a consequence, invalidations now
work independently of Sync's state, and as a nice side-effect, a lot of
plumbing code can be removed.

Bug: 919788
Change-Id: I773e60c7be9a3a979d51412548aa6474001dd7b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2649932
Commit-Queue: Marc Treib <treib@chromium.org>
Reviewed-by: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847574}
GitOrigin-RevId: 72aecf35e2ac0207b012f76d27e6a52d8ceec659