[iOS] Add support for executing JavaScript in WKContentWorlds
Extend JavaScript utils to allow for script execution in specific
WKContentWorlds and frames specified by WKFrameInfo instances.

Previously, executing JavaScript relied on a windowID check to ensure
that the script gets executed on the same page which the native code
intended. This is not necessary for isolated worlds as long as a
WKFrameInfo instance is provided, because the -[WKWebView
evaluateJavaScript:inFrame:inContentWorld:completionHandler:] API
guarantees the script will only be executed in the given frame. (And
similarly to windowIDs, WKFrameInfo instances are recreated on page
navigation.)

Bug: 1042335
Change-Id: Ia1dd7c6e4736337b4fb3b202189d9ac4dfa2b3b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2644152
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#847755}
GitOrigin-RevId: 9a5d768c797ba2a48d8fca94dae450b732999621