Put primary account in INVALID_CREDENTIALS state if missing
This CL puts the primary account in INVALID_CREDENTIALS state if we fail
to load it from the device. This ensures that we load the primary account
in an permanent auth error state instead of it not appearing at all
in the IdentityManager in case there is any error loading the account
information from disk when the IdentityManager initializes.

Bug: 897470

Change-Id: I2f4f52bea8d2d04cf66cbb8b2dabcf5eb8810666
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2637478
Reviewed-by: John Wu <jzw@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#848260}
GitOrigin-RevId: 914ccb70c7fd2baef190fe3056d8acda4f4c1a96