[iOS] Don't ask for reauth if there was no incognito tabs
This CL changes the behaviour of the reauth for incognito tabs, acting
as if the user reauthed since last behaviour if there was no incognito
tabs.
It avoids having the reauth screen showing while the app is still in
foreground (for example with multiwindow).

Fixed: 1169632
Change-Id: I7cb77970a86b7392b3453b8936c5c780cf1dfd5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2659038
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#848553}
GitOrigin-RevId: 6db5dd0ec25cfd006eacedbb29741865a31971ce