[iOS] Ensures that sticky omnibox works when navigating back to NTP
Fixes issue where navigating back to the NTP doesn't stick the omnibox
to the top of the screen when it should.

Lays out content suggestions after the offset has been set to ensure
that the fake omnibox is created. Then, updates the omnibox dimensions
when the NTP appears to make sure they're correct.

Renames updateFakeOminboxOnCollectionScroll to
updateFakeOmniboxForScrollPosition since it's not always called in
response to a scroll event.

Bug: 1172528
Change-Id: I1b5ae3919c6f587a65b6592d3669965bef7f01c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2660815
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#849737}
GitOrigin-RevId: 9143cce3527f93105dbf2658b9b74b39512c5d26