[iOS] Add JavaScriptFeature::FeatureScript::ReinjectionBehavior
In the past, all feature JavaScript was wrapped in a script check to
ensure it is only executed once. This was implemented in order to
prevent bugs caused by WebKit re-injecting WKUserScripts when the
document is re-created, but the window JavaScript object remains the
same.

However, by preventing the JavaScript from being executed after document
re-creation, event listeners are lost and not re-added. Features need
control over this behavior, so it is being exposed through the
JavaScriptFeature class.

While there may be bugs in existing JavaScript features due to a lack of
re-injection, they will soon  be migrated to JavaScriptFeature
instances. During migration, features will be audited to determine if
they should have their script re-injected or not after the document has
been re-created.

Bug: 1042335
Change-Id: Ie24e891751fd00401e57859bf8dba50174b44d90
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2666473
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#849847}
GitOrigin-RevId: d32ef210fde6e9d08bea6795e479f43318d9a21a