[iOS][Thumb Strip] Dismiss keyboard when the thumb strip is Revealed
When the thumb strip transitions to Revealed state, the keyboard
should be closed if it is open onscreen. To facilitate this, the
-willAnimateViewReveal method now has both the current and the next
state instead of just the current state. This is necessary because when
the current state is Peeked, the transition could proceed in either
direction.

Bug: 1094335
Change-Id: If36fc6717c538741c1f66e486c7a5d6a6d34d78e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2662992
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850070}
GitOrigin-RevId: 415ad7c365ec9a21cef4b1663c2fb65919345a5a