[ios] Configure new tabs to inherit opener
Tabs open with "New Tab" button or "⌘+T" keyboard shortcut should
inherit the currently active tab as opener in order to switch the
focus back to the original tab when they are closed (this is also
the behaviour on desktop).

When switching tab, reset the opener of those "New Tab" tabs and
clear the openers when opening them to avoid having a model that
is too complex to understand by user (this is also the behaviour
on desktop).

Bug: 859817
Change-Id: Iee0963af5ffbffe0a21505297c555f27ec422cd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2582086
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850122}
GitOrigin-RevId: 8229bdd327cf1c72e309510bc962f1dd59a3fbf1