Renaming CompromisedCredentials and related structures
This is the first change in the sequence to rename
CompromisedCredentials into InsecureCredential. The reason for renaming
is changed table name. Type alias is used to ensure smooth renaming
process and reasonable CL sizes.

In this CL CompromiseType is renamed to InsecureType and
RemoveCompromisedCredentialsReason to RemoveInsecureCredentialsReason.

Bug: 1137775
Change-Id: I35766404a9f2a925e18486eca54357f7912c8814
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2621059
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Reviewed-by: Jan Wilken Dörrie <jdoerrie@chromium.org>
Commit-Queue: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/master@{#850203}
GitOrigin-RevId: 8f9243d24efbb41aded2e32fdb25f20b05634a0b