breadcrumbs: create breadcrumbs component
This change is the first step toward bringing breadcrumbs (currently
iOS-only) to desktop by making it a layered component. To this end,
this change:
* creates the components/breadcrumbs directory
* establishes DEPS rules for the new directory and those that depend
  on it
* moves crash_reporter_breadcrumb_constants.h into the new directory

crash_reporter_breadcrumb_constants.h was chosen as the first file to
move because it has no dependencies of its own. This allows this
change to be as small as possible while establishing the basic
directory structure needed for future componentization.

Unit tests will be migrated later, after the bulk of the migration is
finished.

Bug: 1163762
Change-Id: Id9f981e120c6d3d00095ef9f387145a385043a78
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2658581
Commit-Queue: Jesse McKenna <jessemckenna@google.com>
Reviewed-by: Cait Phillips <caitkp@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850399}
GitOrigin-RevId: c3416761ea6c76573694c40a84963ca75f96d77b