[iOS][Passwords] Submission detection on change password fields clearing
Unlike login forms, change password forms are often kept on page after
successful submission. This CL allows to detect submission for change
password forms that are cleared field by field by setting an empty value
for all of them.

This change requires an additional mutation observer, but it's initiated
only when the user is manually filling or typing into a password field,
therefore it is not active at all in many cases and when it's active it
observes only one field's attributes.

Base performance was tested based on the following benchmark averaged
over 10 runs on a device: http://browserbench.org/Speedometer2.0/
Without this change:
58.73 +/- 0.53
With this change:
58.50 +/- 0.54

Change-Id: Ia0808c6220f5aa99bf21f67dfa7e477be90b4c10
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2657255
Commit-Queue: Maria Kazinova <kazinova@google.com>
Reviewed-by: Vadym Doroshenko  <dvadym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#850647}
GitOrigin-RevId: 8b63cf72d45f81c69c27a7332c19cc2f3e8fa46a