[iOS] Display the sign-in infobar on navigation redirects.
When a user navigates to Google subdomains via the Omnibox, Chrome
performs an automatic redirect to AuthLogin before continuing to the
domain. In this case the infobar is lost before we display it to the
user. This patch ensures we display the infobar regardless of the
navigation type.

Bug: 1172240
Change-Id: If5142e38c112f52f9dcffe8c029bcb01a25afa1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2675926
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851078}
GitOrigin-RevId: e1ca5960c0cd5b0f19f1ce199e80fa0c00452c46