[iOS][Thumb Strip] Handle calls to -setState: during transition finish
Before, calling setState while a transition is finishing would cause
the layout switcher state to get offset from the overall view reveal
state because we can't start a new transition when a transition is
finishing.

Now, when we programmatically set the state, if the layout transition
is finishing, we cache the state change request and replay it after the
transition finishes. -setState has also been renamed to -setNextState
and explicitly calls out that this transition may not happen
immediately.

Bug: 1173696
Change-Id: If9358add10b47541255167ba0e4b00fc270b34d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2672784
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851231}
GitOrigin-RevId: 436a58753aacabdb7cf7b5fac58b4b716aa48c61