[ios] Use a fixed session identifier if multiple scenes is not supported
For devices that do not support multiple scenes (devices running version
of iOS older than 13.0, phones or old iPads) use a constant identifier
instead of `nil` or `-persistentIdentifier` of UISceneSession (when iOS
version is 13.0).

The reason for using a fixed identifier is that if the device runs on
iOS 13.0 or higher, the swipe gesture is sometimes identified as either
a gesture to kill the app or to close the current window. If the former
happen, on next startup the value of `-persistentIdentifier` will be
different, leading to loss of the session (tabs) and snapshots.

To fix this there are two possible strategy, of either having separate
code path for devices that support multiple scenes and for the others,
or to have a single code path but with a fixed identifier for devices
that do not support multiple windows.

The latter option (fixed identifier) is simpler as it can be limited to
a single location (`-sceneSessionID` in SceneState) if we ensure that
the value can never be the one returned by `-persistentIdentifier`. As
this value is an UUID, it is relatively easy to select a constant that
will not conflict (selected constant is {SyntheticIdentifier}).

To avoid losing data when the user update Chrome, some migration code
is also present to move the session (tabs) and snapshots from the old
location to the new one if needed. Determining the previous location
for those files can be a bit tricky as the location was changed when
multi-window support was added (and enabled in M87).

Bug: 1165798, 1117317, 1138219
Change-Id: I6bbd66651ff9beb5c99138afb57225501a922950
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2644813
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Auto-Submit: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#851707}
GitOrigin-RevId: 92b4838cda71a9a0ec6199dd8d436018e764721f