[iOS]🏢 Open new window in the forced mode
If the incognito mode is forced on or off by a policy, this policy
should be followed when opening a new window.
This CL makes sure that the newly opened window is opening in the
correct mode.

It also updates the bookmark context menu code.

Fixed: 1175063
Change-Id: I3e5919d1c45ff6b0912d501aa5b18f461edc26f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2675914
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#852257}
GitOrigin-RevId: d921907c11182be3e914a7b2d2130ff9c287c0e7