[ZPS] Fix how authentication status is determined in local history zps
Implementations of AutocompleteProviderClient's IsAuthenticated() invoke
IdentityManager::HasPrimaryAccount(signin::ConsentLevel::kSync) which
means AutocompleteProviderClient's IsAuthenticated() actually returns
whether the user is signed in and has granted sync consent. This
contradicts the documentation which suggests this method should return
whether the user is simply signed in. AutocompleteProviderClient provides
another method, IsSyncActive(), that returns whether:
1) the user granted sync consent.
2) sync is enabled (e.g. not disabled by policy)
3) sync is working (not in error)

LocalHistoryZeroSuggestProvider relies on AutocompleteProviderClient's
IsAuthenticated() to decide whether to rank local history zps above or
below remote zps suggestions. If the user is signed in but has not
granted sync consent, this method would return false which could result
in the local zps to be placed above remote pSuggest since the cookie jar
does contain the signed in account and from the server's perspective the
user is signed in. This CL changes how the authentication status is
determined in LocalHistoryZeroSuggestProvider by ignoring the sync
consent so that local history zps would never be ranked above remote zps
for signed in users regardless of the sync status.

Fixed: 1173267
Change-Id: I4eb8bff6b8714eda703a6a5ee8a57ec1724ac3ae
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2664908
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#852446}
GitOrigin-RevId: 549e40bc495aac3c5525b8d82896165d266c205c