[ios][PhishGuard] Port rest of ChromePasswordProtectionService
This CL continues to port over PhishGuard functionality to iOS.

Code in this CL can be compared to the /content counterpart
(from which it was ported) in //chrome/browser/safe_browsing/
chrome_password_protection_service.cc.

Note: password_protection_service.mm was deleted since all of the
implementation is in a derived class.

Bug: 1147967
Change-Id: I18ef13c0f69eaa5ebfdda67c7b566a1e9720255d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2683544
Commit-Queue: edchin <edchin@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#853031}
GitOrigin-RevId: 3ef8ca5f376407ac3e14984ca6e4ad23d33f0958