[iOS][Thumb Strip] Close Thumb Strip on navigation in active tab
As described in the bug, navigation in the active tab should close the
thumb strip, as that represents a signal from the user that they want
to interact with the current tab. Currently, there is no check for if
the navigation is user-initiated (vs e.g. via Javascript). This is for
simplicity. If this ends up being an issue, this check can be added
later.

Fixed: 1175784
Change-Id: Ifee371221a2f9cd997cc0ceda2ed7cf17954d672
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2682343
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#853151}
GitOrigin-RevId: 9222d1ea7ebb7cc15d883bda8c6a1c80dd9aee2d