[ios] Add main/incognito URL drag and drop multiwindow eg test
This should fulfil TC 10/10 on the list of manual test cases.

Changed XCUI helpers to match Cell and windows only.
(There was an issue with multiple matches in window 1 with
one tab. The number of tabs button [1] and the window both
have "1" as identifier)

Bug: 1169687
Change-Id: I39fbaf90e396419bb1e2ae24e808d4a0b7f9d864
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2689763
Reviewed-by: Mike Pinkerton <pinkerton@chromium.org>
Commit-Queue: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#853168}
GitOrigin-RevId: bbc0264ce153d31bc73a1e076507b43126ead40b