[iOS][Thumb Strip] Close thumb strip when focusing omnibox
This CL adds a new command, ThumbStripCommands, which is handled by the
ThumbStripCoordinator. This allows arbitrary objects to close the thumb
strip without needing a reference to the actual PanHandler.

One small change is that the bug says the thumb strip should be closed
when the omnibox is tapped, but I implemented this closing whenever the
omnibox is focused for any reason.

Fixed: 1171366
Change-Id: Id77c26f7e10db78981dd4f96a2fb70926a26f474
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2684576
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#854263}
GitOrigin-RevId: 8fda06a3cedeb0b3ecedebaa0baf62fff5e231b3