[iOS Enterprise] Add managed setting UI for MetricsReportingEnabled
Disables the "Help improve Chrome" toggle in the Chrome iOS settings if
the MetricsReportingEnabled policy is set to false. Note that the
MetricsReportingEnabled policy cannot be set to true at the
POLICY_LEVEL_MANDATORY level, so if the policy is set to true, it will
remain at the POLICY_LEVEL_RECOMMENDED level and the user will be able
to change the pref in the Chrome settings.

The policy itself was already supported, this CL only fixes the UI so
it correctly supports the policy.

Bug: 1085053
Change-Id: I8f93737e2e85a67602fb7f50400497b9c0d6345e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2692651
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#854349}
GitOrigin-RevId: f61738121c8f4aeaa0af264c20f83debc3b13c50