[iOS][SH-iOS] Use Canonical URL As Link-To-Text Base URL
The JavaScript call to generate a link-to-text will also return, if
present, the page's canonical URL. Then, when on HTTPS pages, Bling will
use this URL as base for generating the final link-to-text.

This change brings link-to-text on-par with the existing Omnibox share
functionality, which also uses the canonical URL when present (and on
HTTPS pages).

Fixed: 1179355
Change-Id: I46834d5288f5db0cfb9567f9d4d13f1b2de3404f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2703654
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855298}
GitOrigin-RevId: 3cc9671da76c315e4d03c8198b330a9f8dc9e0b2