[iOS Enterprise] BrowserSignin: Crash if sign-in isn't allowed
Adds a CHECK in the AuthenticationService at the beginning of sign-in
that ensure sign-in is allowed before proceeding. This ensures that the
BrowserSignin policy is respected when set to 0 (sign-in disabled) even
if some sign-in entry points in the UI are still accessible.

Note that the sign-in process would have crashed anyway due to the CHECK
on line 341, because the IdentityManager won't set a primary account if
kSigninAllowed is false. This CL simply makes a more explicit CHECK.

Bug: 1155745
Change-Id: I8ccb7407acd1947bd648917148628466b975495b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2696422
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855498}
GitOrigin-RevId: 9c363142c4952655580cdc966c34fe50482d0b74