[ios]🏢 Consider the incognito policy when opening chrome links
Currently, whatever link we open from the chrome menus (e.g., from the
browser menu), will be opened in a regular tab. When using the Force
Incognito policy, we want all tabs to be opened in incognito.

In this CL, we propose a solution that will change the application mode
to incognito for the links opened from the browser menus when the Force
Incognito policy is active.

The logic we have implemented also handles the case where incognito is
disabled by the policy. Technically, we wouldn't need that because the
default application mode is REGULAR, but we want to make sure that the
mode is coherent with the incognito policy.

Tested these scenarios (see bug for more details):
 * (Only Incog) Open "Learn More" from the browser menu.
 * (Only Incog) Open "Learn More" from the enterprise help popup in
   Settings.
 * (Only Incog) Open "Privacy Notice" from the enterprise help popup in
   Settings.
 * (Only Incog) Do all the steps above with multi-windows on iPad
 * (Only Reg) Open "Learn More" from the browser menu.
 * (Only Reg) Open "Learn More" from the enterprise help popup in
   Settings
 * (Only Reg) Open "Privacy Notice" from the enterprise help popup in
   Settings.
 * (Regular Mode) Go through the scenarios without any special incognito
   policy.

Bug: 1179401
Change-Id: I2d7f7f66176a4d9e3fce54e8025c216d398c7325
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2701952
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Cr-Commit-Position: refs/heads/master@{#856446}
GitOrigin-RevId: 9bd0872cc344aea11dd2df21048450ede99c416e