Convert console.js to JavaScriptFeature
Logic previously in JavaScriptConsoleTabHelper is migrated to
JavaScriptConsoleFeature.

The change from messages coming from the WebState level to the
BrowserState allows inspect_ui.mm to be cleaned up as it no longer needs
to add listeners to every BrowserList and WebState. However, this makes
is less clear to the the inspect UI when WebStates are closed, so
messages are no longer removed from the chrome://inspect page when the
WebState is closed. The corresponding test has been removed.

Additionally, the log level and message types in
JavaScriptConsoleMessage were updated to NSString to prevent unnecessary
conversions. (As WKScriptMessage provides the values as NSString
directly to JavaScriptFeature::ScriptMessageReceived.)

Bug: 1175932
Change-Id: I7c29042ebaed05ab49dbc2e905e0fd12b005e00a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2689824
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#856517}
GitOrigin-RevId: 1197fcebee6a737cf6d26d4207bbb62062712411