[iOS][Settings] Fixes clear browsing data link
This CL fixes the 'Learn more’ link.
It corrects the method name so that the delegate is correctly notified.
It also adds new EG tests and a new action 'TapAtPointPercentage' used
to press the top right corner of the selected element when the link is
displayed on 2 lines.

Bug: 1179227
Change-Id: Ib483014da4ef3dd665926df94cef23cdc1490b98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2713142
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#857682}
GitOrigin-RevId: 4eb5f3320c55f103bbd372797ee9184ad16e402e