[ios]🏢 Handle external intent in unexpected mode
Open the external intent in a valid mode according to the value of
IncognitoModeAvailability policy.

If users intend to open a new search when incognito is forced, open
in incognito mode, otherwise open in regular mode by default.

If user intend to open an url in regular tab when regular tab is
disabled, open chrome in incognito mode without open the url,
instead, show a toast telling that the regular tab is disabled.
And vise versa.

When long-press the chrome icon, both "New Search" and "New
Incognito Search" options are available. If users click "New
Search" but only incognito mode is available, open an incognito
search with a toast, and vise versa.

Bug: 1169898
Change-Id: Ia4f0d26619a35c38a0fafaf41ebd8afb211e300f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2645449
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#857842}
GitOrigin-RevId: 6309828eab4fd5d99e61e0f074f852c95c48287b