Mark scoped_refptr with [[clang::trivial_abi]].
We already use the experimental libc++ ABI, which does the same to
std::unique_ptr and std::shared_ptr.

This means that files that don't build with the Windows ABI may
need additional includes (because the destructor runs in the callee's
frame) and the order of construction and destruction of arguments may
not be well-nested. In some cases this also surfaces UAF bugs that
were previously masked by optimization.

This CL includes the additional needed includes, and the test fixes
have been committed separately.

This allows scoped_refptr and trivial wrappers of it to be passed in
registers instead of on the stack, which leads to about 140 kB of APK
size reduction on Android. It may also have a slight performance
advantage, though this hasn't been validated.

Bug: 1019548
Change-Id: I01ca86d2cabb54aed9c51438732506fa87d17a63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2714247
Reviewed-by: danakj <danakj@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Owners-Override: danakj <danakj@chromium.org>
Commit-Queue: Jeremy Roman <jbroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#858306}
GitOrigin-RevId: 7279bdd4185a33855d6a0861ae73af6bea0b71cc