[ios] Attempt at more deflaking of testOpenInPDF|PNG
Still at around 30%, and it seems from the log and images,
that the after the 'open in' button is tapped, there's still
a delay before the dialog appears, and we fail to find
the filename during that delay.

Part of the effort to lower flaky rates in try bots.

Change-Id: I7035bb98fbe95c35aadc8de9ad790b1bffc64d5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2725515
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: David Jean <djean@chromium.org>
Cr-Commit-Position: refs/heads/master@{#858573}
GitOrigin-RevId: 567f1be7c4f32443e845297bd36e9a9bd215ce90