[iOS][Settings] Fixes clear browsing data link
Reland: https://crrev.com/c/2720105
'Learn more' isn't always visible depending on the device, sometimes
we have to scroll to make it appear.

This CL fixes the 'Learn more’ link.
It corrects the method name so that the delegate is correctly notified.
It also adds new EG tests and a new action 'TapAtPointPercentage' used
to press the top right corner of the selected element when the link is
displayed on 2 lines.

Bug: 1179227
Change-Id: I4c4cb1539ce28e85ed168e55187e61b0e80fc8e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2726237
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Ewann Pellé <ewannpv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#858575}
GitOrigin-RevId: da3b228f9bf637520c97e050490b66975685d3bb