[iOS] Allow BVC content to overflow outside the BVC view
This CL removes clipping the BVC's content to its bounds. This allows
content to overflow outside the BVC's view, which is useful when the
BVC's view gets translated down due to Thumb Strip.

I see that this line was added to fix bug 809103. I tested the change
on iPhones on iOS 12.4, 13.5, 14.2 and the bug does not reappear.

Fixed: 1171436
Change-Id: If9aa74eafc75a78433b969e6a1b954aaa5b3771a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2726704
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#858585}
GitOrigin-RevId: ad6d08a847e15015615a8ecfdce571c078e8e5cd