[iOS] CWTChromeDriver: Print a heartbeat message every 30 seconds
By default, xcodebuild_runner.LaunchCommand times out and kills
the app after 180 seconds with no new output printed to stdout
or stderr, since for typical tests, having no new output after a
long period of time implies that the app is hung.

However, for fuzzing, CWTChromeDriver isn't running tests that
produce output in the usual way, so it's common for it to be
running for long periods of time without anything printed to
stdout.

To prevent CWTChromeDriver from getting killed, this CL adds a
heartbeat message that is printed to stderr every 30 seconds.

Change-Id: I7c3f1f1703b9035a3969ada59f6f8009f5e3a6b1
Bug: 1158540
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2645604
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859068}
GitOrigin-RevId: afa52118f28b0ec1392ac076fc6f4d1509729ff0