[ios] Generate a project using Xcode New Build System
The Legacy Build Systems appears to be deprecated in Xcode. As the
project is now generated with all the information required to build
and run using gn+ninja in Xcode using the New Build System, change
ios/build/tools/setup-gn.py to generate a project using it.

Bug: 1177345
Change-Id: I6ebe2d0f126d4045f0986b7c373fdbb5eaa0aab0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2690567
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#859372}
GitOrigin-RevId: a7de205912cf5994da77cb067d1aba1713fb75c8