[iOS][Thumb Strip] Ensure heights are updated correctly on rotation
The root cause of the bug was that the rotation was occuring when the
BVC was offscreen (because the last tab was closed, so it was never
alerted to the size change). Now, this job is handled by the bvc
container, which is always onscreen and always gets notified.

Fixed: 1184101
Change-Id: I68c02e9bc4ff7fac00f45704f0cbbe2338a91a26
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2732428
Auto-Submit: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#859506}
GitOrigin-RevId: eec8844cd1392d5d838892dc0c16a41b1d0749a7