[iOS] Add content world JS support to FakeWebFrames
The CallJavaScriptFunction methods which accept a content world
need to be available to WebFrameImpl and FakeWebFrame, but the content
world is not directly exposed outside of |//ios/web|. In order to
support this, the shared APIs are moved to a new |WebFrameInternal|
header.

This allows the application to continue to work in the same manner
without direct casting to WebFrameImpl (which fails during testing when
the WebFrame is actually a FakeWebFrame).

This change specifically supports the conversion of Find in Page to
JavaScript feature tracked in crbug.com/1175919, but is purposefully
generic enough to continue the use of FakeWebFrames with features
moved to the JavaScriptFeature architecture.

Change-Id: I9083a18d3efd15a199b20adef6434c911e25a90d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2732332
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860125}
GitOrigin-RevId: b74ce54c11b41a4833bc27494e25ac291423a197