[ios] Add ParseStringWith{Links,Tags}
Add new functions ParseStringWith{Links,Tags} that parses a string
looking for pairs of begin/end tags, creating a string without the
tags and a list of ranges pointing to the substrings with the
content found between the pairs.

Reimplement ParseStringWith{Link,Tag} using the new function. Change
the prototype of the function to return a structure containing the
string and the range. Change the functions to assert that there was
at most one begin/end tags pair.

Bug: 1181219
Change-Id: Ie70518f05829d2be3c0e56db2b16cb697404379a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2735817
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860188}
GitOrigin-RevId: 3748b196b9f18e86c9dd9af9681ee3da2d7f0e8a