Check SafeBrowsingAllowlistDomains policy on subresources.
Pass in a list of allowlist domains to the URL checker delegate for
subresources check initiated from the renderer. Since the callback
binding is repeated, the allowlist domains are copied each time. To make
sure it is not too expensive to copy the list, also add a metric to
measure the total size of the list.

Also add a browser test to capture the failure. The test failed before
the patch and succeeded after the patch.

Metrics added:
SafeBrowsing.Policy.AllowlistDomainsTotalSize

Bug: 1179276
Change-Id: Ic33035668d4c3d9ac77b9bf9de9b961a7009f198
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2718546
Commit-Queue: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Tim Volodine <timvolodine@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860498}
GitOrigin-RevId: 03180c4aa8c39c4fbe19103676f50c0ed7ebbb09