[iOS] Stop sending location with omnibox query
This CL stop sending the location with the omnibox query.
The workflow to request the permission to use the location isn't
altered to make sure we are able to request it for the pages requesting
the location from the user.

Bug: 1178627
Change-Id: If0ef705542ccf9b081adfc81559f6b5bd0275834
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2695382
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860849}
GitOrigin-RevId: f612c814ae587c749c0947de96c2b7d42921509f