Shrink Google logo for Start Surface
This change gives the doodle and logo a shrunk height for the Start
Surface. It also moves the avatar image down next to the shrunk image
in portrait mode for iphone. Upon rotation to landscape, the avatar
moves back to its default location in the top right.

Screenshots:
https://drive.google.com/file/d/1KwI4LrAkBlIWc7Sf3oh96dfs7ZeowqOw/view?usp=sharing
https://drive.google.com/file/d/1qcwS-sjkeW9jUtX1oCiRAvXNFkE3N3C1/view?usp=sharing

Bug: 1170491
Change-Id: I6c3c7e44b898a0a09552719fdf89afcdad6789e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2733133
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#860883}
GitOrigin-RevId: e0493deff8a75e0930cec76a4e9673f1adc8ca59