[iOS] Unfocus omnibox when changing feed visibility
It is possible to access the feed header menu with the omnibox focused,
allowing users to turn off the feed. This would never trigger the
cancelling of the omnibox, leaving it there until the next time it's
used.

This CL cancels the omnibox editing when changing the feed visibility.

Bug: 1178471, 1179230
Change-Id: Ie2598638afb6fada5917d6770f1c13e39b5d4ea6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2745277
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#861594}
GitOrigin-RevId: d76cdc2023604d43e708b9b6665e66b0b36442aa