[iOS] Fix top scroll position when rotating
Rotating the device to landscape removes the fake omnibox, reducing the
height of the content suggestions. Since the NTP's top scroll position
is dependent on the content suggestions height, rotating from
landscape->portrait would mess up the top scroll position.

This CL checks if the content suggestions height was increased when
rotating, and if the NTP was scrolled to the top. If these conditions
are met, then the scroll position should be set to ensure that the NTP
remains at the top position.

Also removes unnecessary scroll delegate functions.

Bug: 1174052
Change-Id: Ie892a58f6fadb73fbb3471162e9d03b850ac4f7d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2745575
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Cr-Commit-Position: refs/heads/master@{#862055}
GitOrigin-RevId: f943fddebf2a1c49e508944e0f0ec3c9d19f6a8a