Extensions: Plumb whether request was blocked by an extension.
Currently navigation requests blocked with net::ERR_BLOCKED_BY_CLIENT
error code are all incorrectly reported as being blocked by extensions.
This CL plumbs whether the request was blocked by an extension using the
opaque "extended_error_code" field of URLLoaderCompletionStatus. To
avoid layering violations, blink::ResourceRequestBlockedReason is
extended by embedders.

This helps in showing a more correct error message to the user when a
non-extensions client blocked a main-frame request using
net::ERR_BLOCKED_BY_CLIENT. The "extended_error_code" can further be
used by other net::ERR_BLOCKED_BY_CLIENT clients to improve the error
message shown to the user.

This information can also be used by Devtools in a future CL to show
whether a request was blocked by an extension.

Also, the ResourceRequestBlockedReason::kCollapsedByClient is now
removed and is replaced with a separate field. This was necessary since
the "extended_error_code" needed to be kExtension in this case. It also
makes semantic sense since "collapsing" is an action to take and not a
reason.

BUG=649869

Change-Id: Iab66268ef9110b73301cfeaa79724f00036594af
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2721253
Commit-Queue: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Edward Jung (EMEA) <edwardjung@chromium.org>
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#862652}
GitOrigin-RevId: 3de98e520cea74e1f1bdc8a112bb66a8f2a19541