Sync internals: Autogenerate grd file
Autogenerate the grd file for sync internals. Update the data sources
(both iOS and standard webui) to add the resources accordingly. Note:
This registers the CSS files individually, instead of inlining them
with flattenhtml, as flattenhtml is no longer recommended for WebUI.

Adding dependency for building the iOS data source, so that it can
use the new ResourcePath struct that is used in Grit resource maps.

Note to iOS sheriff: There is a follow-on internal CL to this to fix
internal iOS builds, created by treib@ (see comment thread for full
details); please consider landing that CL instead of reverting this,
if you see breakages on internal bots.

Bug: 1173575
Change-Id: I046134fc75ea410294417ccdbac9c65027d9cc0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2729735
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#863177}
GitOrigin-RevId: cc2b166983075dfc3f8e7d99c91a229fd446803a