Listen to FaviconDriverObserver to update Return to Recent Tab tile
Upon cold start, the favicon associated with the current page in
a WebState is not available and has to be fetched from the db. Thus,
the favicon will not be available when the Return to Recent Tab tile
is first configured. A fetch needs to be executed and a Start Surface
object needs to listen to FaviconDriverObserver to received the fetched
favicon.

StartSurfaceRecentTabBrowserAgent is a good fit for this observation
to be able to then send the favicon to ContentSuggestionsMediator.
The name of the browser agent observer is changed to reflect this new
additional purpose.

Bug: 1173160
Change-Id: I061e7696ac703eaa718ebfee2b225975f16ebfe3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2765369
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864041}
GitOrigin-RevId: 387c02502aec2efeb6e42784018ad2144700e86d