[Lookalikes] Add TriggeredByInitialUrl to Lookalikes UKM
This CL adds a boolean to the UKM recorded for lookalikes that indicates
whether the URL that triggered the heuristic was the first URL in the
redirect chain (initial_url) or the final URL in the chain
(committed_url). Without this change, we just made educated guesses.
Privacy approved this addition in go/lookalike-ukm?disco=AAAALpzpUVk .

Presently, iOS checks ALL urls in the chain, not just first and last.
Rather than solve that bug here, this CL just sets TriggeredByInitialUrl
to false always on iOS and adds a TODO to fix that in the future.

Bug: 1133598
Change-Id: I2a6c4854e3d6f28e19bb6b03c6a6187d7c49e7e9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2765941
Reviewed-by: Livvie Lin <livvielin@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Joe DeBlasio <jdeblasio@chromium.org>
Cr-Commit-Position: refs/heads/master@{#864260}
GitOrigin-RevId: 1ca12e3313803a326308936ffd7b9f088c11e1b1