[iOS] Fix crash when appending tab to destroyed tab
This CL prevents a crash which would happen when a tab is opened in
background and positioned after the current tab, and the current tab
has been destroyed.
On iPhone, opening a tab in background in asynchronous, there is enough
time to close the current tab. This is destroying the WebState, but
there is still a pointer to it, leading to a crash.

Fixed: 1191312
Change-Id: Ic1c0ae54f41960a071e3590827eef70bc6602eb1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2784862
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: David Jean <djean@chromium.org>
Reviewed-by: David Jean <djean@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#866161}
GitOrigin-RevId: 504b0033fb9e8789911ce9df65fc610183dffc13