[iOS] Exit fullscreen when sharing the page
In some cases (for example when the Share Link is failing) the page can
be shared while being in fullscreen.
On iPad this leads to a broken UI where the share menu is pointing to
where the Share button should be if the toolbar is present.
Exit fullscreen in all cases to avoid this. Exiting on iPhone is also
good as it is displaying the omnibox, making sure the user knows which
page they are sharing.

Bug: 1147507
Change-Id: Ia331dd304830d0340e6539fa6c2fadf85f272e70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2739722
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#866536}
GitOrigin-RevId: 881b738fbbb658afa4f3f6839281ec75ed171e04