Sync internals: Fix some issues with handling empty node lists
- Empty lists cause a console error due to bad handling (all platforms).
  Fix this by checking the returned value.
- iOS has a separate issue due to using custom message handler
  implementation; need to add base::Value(true) to indicate that the
  callback was resolved correctly. Also simplifying the logic a bit
  to remove a call to Clone().

Note: Tested iOS change by copying this code over to the message
handler for other platforms. Was able to repro iOS issue reported on
the bug in this way, and confirmed the change fixed the issue before
copying the code back to the iOS handler and modifying to work with the
iOS version of CallJavascriptFunction().

Bug: 1191388
Change-Id: I18ab0be1d7972d6954bd8eecac31ff38122ac9a7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2785997
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Cr-Commit-Position: refs/heads/master@{#867365}
GitOrigin-RevId: 41e2f99f5089b916054c220958d178ee602f97ac