Restore groups as a single entry in 'recently closed'
Currently, when a user right-clicks on a group header and chooses "Close
Group", each tab in the group gets is own entry in TabRestore and there
is no way to restore the group as a whole.

This CL creates a new TabRestore entry type for tab groups. Then, when a
user right-clicks on a group header and chooses "Close Group", the group
entry is stored as a single item. The group will then restore as a unit,
similar to windows.

Individual tabs that are part of a group can still be closed
independently, resulting in separate tab entries. When those tabs are
restored, the group will still be recreated in the tabstrip. If the
group itself is also closed, the individual tabs and group will merge
together in the tabstrip at the time of restore.

This CL only adds group entries to the app menu. A followup CL also
adds group entries to the Mac and Linux 'recently closed' menus.

App menu screenshot:
https://bugs.chromium.org/p/chromium/issues/detail?id=1078957#c18

Skipping screenshot check for IDS_RESTORE_GROUP because IDS_RESTORE_TAB
and IDS_RESTORE_WINDOW should share the same screenshot. But screenshots
are added for other UI.

Bug: 1078957
Change-Id: I16c82bb688536758e69e0c0e59beeabdb08f20bf
Skip-Translation-Screenshots-Check: True
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2574810
Reviewed-by: Jesse Doherty <jwd@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Commit-Queue: Connie Wan <connily@chromium.org>
Cr-Commit-Position: refs/heads/master@{#868650}
GitOrigin-RevId: 50802788094d7cf18b4682924e7249a6449b3238