[ios] Set metadata to report correct version when using GTM
Some internal dependencies of Chrome on iOS uses GTMSessionFetcher
that reports an incorrect version number in the UserAgent because
it uses the value from CFBundleShortVersionString (this number has
to be composed of one to three period-separated integers which is
not space to store the full Chromium version).

With https://github.com/google/gtm-session-fetcher/pull/227 it will
be possible to set GTMUserAgentVersion to be used and override the
version reported. Additionally, set GTMUserAgentID to also report
the application name instead of defaulting to process name.

Bug: 827856
Change-Id: I380d4ce4ea2fd5f81bf9ff80d2cfab44211a5635
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2787086
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#869599}
GitOrigin-RevId: 1a8583d7a852796182f9e19e93d111e822ba8b9b