[SH iOS] Remove async checks in ShouldOffer
This switches from an asynchronous, JavaScript-style approach to
enabling/disabling the Create Link (a.k.a. Link to Text, Shared
Highlighting, etc.) option in the Edit Menu, to one based on examining
a UITextInput to identify properties about the user selection. Our
hypothesis is that the asynchronicity was at the root of a number of
crashes observed in Beta.

Bug: 1187424
Change-Id: Ic924a2b16b9343a06f5753fe32dcb6bb32710f94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2787333
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Cr-Commit-Position: refs/heads/master@{#869614}
GitOrigin-RevId: 7ab345319a0f4aa7ba79b2c8aa12bb93afdb845a