[iOS] Add a fuzzing test runner script
This CL adds a run_fuzz_test.py script that can be used by
Clusterfuzz to load a single test page. This script ensures
that CWTChromeDriver is running, constructs a file URL pointing
to the test page, passes this URL to CWTChromeDriver, waits for
output, and then echoes this output.

This CL also prevents CWTChromeDriver from getting killed by XCTest
when a crash is found. When a crash is found by ASan, a SIGABRT signal
is produced in the app process. However, if the app process dies
because of SIGABRT, XCTest kills the test process as well. To prevent
this from happening, this CL adds a CWTWebDriverAppInterface method to
install a signal handler that catches SIGABRT and then exits normally.

Bug: 1158540
Change-Id: Idbbcf76438a768ad7da85aff299ad14fa70164dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2808553
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#870112}
GitOrigin-RevId: 348dd2b9bc17384117fbb4daba0a36b5f9d3e823