[ios] Stop accessing _browser after -settingsWillBeDismissed
If an authentication is in progress during the app shutdown,
it is possible for the authentication to complete after the
settings have been shutdown but before the controller could
be deallocated (e.g. it may be autoreleased and the poll not
drained yet).

Then -[AccountsTableViewController popViewIfSignedOut] could
be called with a _browser being a dangling pointer leading
to UAF and in the best case a crash during shutdown (it is
probably hard to exploit as it requires the user to kill the
app at a precise moment, so unlikely to be a security risk).

Clear _browser in -settingsWillBeDismissed and ensure that
the code does not use -authService when _browser is null.

Bug: 1192993
Change-Id: I788177beda6b0a5eab64f20e52573305412dd96c
Fixed: 1192993
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2810298
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#870444}
GitOrigin-RevId: 58dbfd8e942d5f605ecd8ebea3f6812faa17e35f