[iOS][Default Promo] Log omnibox actions for non-modal promo
The omnibox actions need to be routed to the non-modal promo scheduler,
so they all need to go through a coordinator/mediator.

Also, because the omnibox popup mediator is constructed during the
process of creating the main browser, the DefaultBrowserSceneAgent must
exist before this. Before, the DefaultBrowserSceneAgent wasn't created
until after the main browser because it requires a command dispatcher
provided in the initializer. Now, the agent is created before the
main browser, but the dispatcher is provided afterwards.

Bug: 1191535
Change-Id: I51d22999aefa04c1be7c1ab3c5e667ec8d60ce70
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2808982
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#870954}
GitOrigin-RevId: 9a636e056d606ef6a17a80def04fbccfceb484d7