[iOS] Split createMainBrowser method in two.
Some AppAgents require a dispatcher at init. -createMainBrowser, as the
name suggests, creates the main Browser, but also it starts the main
coordinator, which basically inits most of the UI.
Some UI needs scene agents.

I think in general scene agents should be created before the UI is
created. They are "root" scene-level objects, so in UI code I'd expect
them to exist. Therefore I'm splitting createMainBrowser in two, which
allows to add code when the main Browser (and its dispatcher) are
created but the UI is not. The existing coordinators that require the
dispatcher are added in this intermediate stage.

Bug: none
Change-Id: Id247c6a9bcd4e2adcf3fd06f27f13226ff47a3b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2814682
Commit-Queue: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Robbie Gibson <rkgibson@google.com>
Cr-Commit-Position: refs/heads/master@{#871005}
GitOrigin-RevId: bc6830864c9776acc5433e6e1e80347be0da6cb5