[iOS] Split the app initialization in pre and post safemode
This is CL is an intermediate step towards the safe mode agent (
https://chromium-review.googlesource.com/c/chromium/src/+/2757684)

This will also unblock the initial layout for the enterprise agent (
https://chromium-review.googlesource.com/c/chromium/src/+/2784892)

We also change didTransitionToInitStage to didTransitionFromInitStage.
We can isolate that change in a standalone CL if needed.

This CL splits the initialization of the browser into 2 steps: (1)
initialize the minimal browser basics to support safe mode, and (2)
complete the initialization of the browser when transitioning out of safe
mode (in both cases, when safemode is or isn't triggered).

We add the extra InitStageBrowserBasic stage to handle the initialization
of browser basics.

Bug: 1196811
Change-Id: I7ddebe857b9fade28fa6ea88d04fe6de6f0a2b83
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2811498
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/master@{#871040}
GitOrigin-RevId: 06a84d4699fe43e47415c39428f70b0528ee7c96