🐻 [Pandora] Create basic MVC module for Sync screen
Create a skeleton of the MVC component with classes as below:
A coordinator as the interface of this screen.
A mediator talk to the model, update the UI and store sync status to
the model.
A view controller that create the UI.
A view controller delegate that handles the user click actions.

Bug: 1186767
Change-Id: I2e4a6f49f98a04ef7a469282da16beacab39565e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2827516
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#873417}
GitOrigin-RevId: fee918ccc01749e5ca29578505f359355e2a7703