[iOS][MICE] Remove the scrollview from being under the navigation bar
In the consistency bottom sheet, the main view of the first panel is
the scroll view (instead of a regular UIView). Therefore, the user
scrolls the view down, the panel content goes under the navigation bar.

To avoid seeing the text under the navigation bar, this patch changes
the main view of the view controller to be the default view (created be
UIViewController), and the scroll view is now a subview of
-UIViewController.view.

iPhone 14:
https://drive.google.com/file/d/1RCRn7AQTmBEjOfMBkDpZrORKm2K6dvtT/view?usp=sharing&resourcekey=0-dJkkyYPBwjF-sNJu12VgYA
iPhone SE:
https://drive.google.com/file/d/18NmQ6fbv7Y4Q2siOE_tA2XyM0BZNddfu/view?usp=sharing&resourcekey=0-0UD4pENnEmjvpUff5m4Bng

Bug: 1151288
Change-Id: I57e77fbd66fe1dbe08ed9f5d690e8d7994f1c539
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2837982
Auto-Submit: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Commit-Queue: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#874262}
GitOrigin-RevId: 577316c2af46e1f4e2724c01aa576dae72b65ce3