[iOS FRE] Add iPad layout constraints to shared FRE screen VC
Adds constraints for the iPad layout in the shared first-run screen
view controller. The only constraints that change are the width of the
action buttons and the width of the scroll content view, which should
take 65% of the view width while in compact width, and the full view
width (minus some margins) while in regular width.

Bug: 1189850
Change-Id: If12fb1a92c46653c348c7742eab4360524eee610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2829072
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#874631}
GitOrigin-RevId: 7752133f140e2323f5e263a7c5d8a2c0de99e30f