[iOS][Default Promo] Add timers for show/dismiss of non-modal promo
This CL adds the initial 3 second timer after an omnibox paste to show
the promo and the timer to dismiss the promo after it has shown for some
time.

It also observes messages and webstate to prevent the promo from showing
if a message is presented or the active webstate changes in between
the triggering event and the actual promo showing.

Fixed: 1191541, 1191542
Change-Id: Id2f79b0f30365fc5b5b8ecbe8bb670d4ed98977c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2835771
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#876124}
GitOrigin-RevId: e11d8317b7e5c3b60b7c6a86b2f96166b46a8025