[ios] Make SpeechInputLocaleConfigImpl testable
Inject the lists of VoiceSearchLanguage and SpeechInputLocaleMatch in
SpeechInputLocaleConfigImpl via its constructor to allow testing the
class independently of the provider that is registered.

Remove SpeechInputLocaleMatchConfig class and change the loading code
to use a free functions (instead of a property on a class implementing
the singleton pattern).

Add unit tests for SpeechInputLocaleConfigImpl.

Bug: none
Change-Id: I8796ecdd8c328cc72c0f9cb0d7acdecebc55c879
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2851160
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#876186}
GitOrigin-RevId: fa664d91ffc14af0161e050b3581de966502edb6