[iOS] Prevent crash from invalid session restore data
This makes the logic for using the session restore API found in
trunk WebKit handle the case where there's an error when
unarchiving session state from NSData.

This also removes a call to |goToBackForwardListItem:| that's not
needed, since the call to |setInteractionState:| already implicitly
performs this navigation.

Bug: 1174560
Change-Id: I0c70ee83727ff1e813b9294ab0d3c5bd8aadfdc5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2850435
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#876558}
GitOrigin-RevId: 298e51b1c64b5c0488d4285e0edbd96e7747cb64