[iOS] Stop asking for permission on omnibox focus
This CL makes sure that we stop asking for the location permission when
the user focuses the omnibox.
It is also no longer asking for permission when loading a google.com
page.

If the user didn't provide an answer to the permission request, the
permission is being asked when a page is requesting the location
through the Geolocation Web API.

Fixed: 1196217, 1201731
Change-Id: I9cc7c0d2018b89fb96c90953deedf517fb89dcb9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2851753
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Auto-Submit: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#877029}
NOKEYCHECK=True
GitOrigin-RevId: 5d3fbe47600b2f30f37523ea75b62f5fa2ce1eef