[ios] Simplify OpenInController implementation
Use free functions to implement the file manipulation operations
instead of methods / static methods which is easier to call from
blocks / callbacks without inadvertently capturing state.

Remove OpenInControllerBridge which is not really needed. It is
simpler to just use a weak pointer to OpenInController captured
by a block.

Add a boolean to record the fact that -disable was called. This
is needed to prevent executing the code in the block mentionned
above when the object has been disabled but not yet deallocated
(e.g. it could be in an autorelease pool).

Use a sequence checker to check callbacks are executed on the
correct sequence instead of checking for a named thread (this
makes code easier to write and test).

Ensure that OpenInController drop its pointer to Browser and to
WebState when disconnected from the UI or when the WebState is
destroyed.

Bug: none
Change-Id: I61d7f7b31704685364c0f747517d3f0ee7dee465
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2854920
Reviewed-by: Roger McFarlane <rogerm@chromium.org>
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#877074}
NOKEYCHECK=True
GitOrigin-RevId: f45caf59d688000a1e5d19b8e8d0daaf49b2e981