[iOS][Mice] Adding spinner loop while sign-in
The spinner is added on top of the "Continue as" button. Also this
button and the account chooser button is disabled when the spinner runs.

The spinner will be used in the following patch:
crrev.com/c/2859844

https://drive.google.com/file/d/1uNnz4ozGuQFWprSMPGGXs9x9djjfsySx/view?usp=sharing&resourcekey=0-j5DlDMoHPTq05l5Yc3dHeA

Bug: 1151288
Change-Id: I50e886fa8b6efdbbe83f8bded44d669fea449080
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2859922
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Auto-Submit: Jérôme <jlebel@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#877531}
NOKEYCHECK=True
GitOrigin-RevId: 2dd6ef8225f529a9c0197bf9add5dd04e6134943