[ios] Listen to WebStateDestroyed in StartSurfaceRecentTabBrowserAgent
Listening to WebStateDestroyed ensures that
favicon::FaviconDriverObserver, which is tied to a WebState, can be
reset in the instance that the WebState is destroyed in a different
logic flow that doesn't execute WebStateDetachedAt(). f.e. if an
existing WebState is replaced.

Fixed: 1204328
Change-Id: I1daaed3e6c013fc7973e70ad0cf4e74c954c1106
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2861061
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Reviewed-by: Ganggui Tang <gogerald@chromium.org>
Cr-Commit-Position: refs/heads/master@{#878023}
NOKEYCHECK=True
GitOrigin-RevId: a2cb24bdfc60a2b1d7d5d8b9a2e064aa1acb2165