[iOS] Fix crash when toggling feed and add tests
Fixes crash when the feed was hidden/disabled by setting the
DiscoverFeedVC to nil when restarting the NTPCoordinator.

Adds tests for when the feed is hidden/shown, and for when the feed
is disabled/enabled. Tests will check that the feed header reacts
properly.

Bug: 1194106
Change-Id: Ibcac163a633749868ed2d04558d8e01515c8fb98
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2860944
Commit-Queue: Adam Trudeau-Arcaro <adamta@google.com>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Cr-Commit-Position: refs/heads/master@{#878550}
NOKEYCHECK=True
GitOrigin-RevId: d43462eb179ca10efe16849d42dcec835b623954