[bookmarks] Fix crash if observer destroys itself
The first event notified in BookmarkModelBridge::BookmarkNodeRemoved()
can have the side effect of deleting the bridge itself. To avoid crashes
in this situation, keep a copy of the weak reference to the target
observer.

It is questionable if the second even should be notified in this
situation, and in this tradeoff, it appears safest to err on the side
of broadcasting both, in case the client code assumes both calls are
always invoked together.

Change-Id: I334d2f4ee9ee6387c041cd921a7ec3dbcca9555c
Bug: 1204155
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2859939
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#879264}
NOKEYCHECK=True
GitOrigin-RevId: 9bacbcbf1e5721f89a625f2e21ad8ebe87f631ca