[iOS][Default Promo] Add timer to prevent promo from showing too often
The current rules are:
- at least 2 weeks between impression of any promo
- max 2 impressions of the non-modal promo.

An "impression" is also defined as any promo where the user interacts
or the promo timesout. This does not include cases where the promo is
dismissed because of some other action (e.g. backgrounding the app or
entering the tab grid).

Fixed: 1205063
Change-Id: Ibb73940786419011ddf43807bbfb805f6a472d97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2867530
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Javier Flores <javierrobles@chromium.org>
Cr-Commit-Position: refs/heads/master@{#879303}
NOKEYCHECK=True
GitOrigin-RevId: 39c39bf723529050c1136ace81dd08a16492147b