[iOS] Remove the iOS 12 block in SceneController#handleExternalIntents
This concerns external intents handling such as open URL on iOS 12.

The launch options will always be nil when handled in
-handleExternalIntents because the launch options are cleared before.

This means that there are no situations where there are launch options
to be handled at that point in the initialization sequence.

The startup parameters are initialized by the intent handlers themselves
from their options.

The startup parameters are handled later by -resumeSessionWithTabOpener.

An important point is that on iOS 12 the startup sequence to handle
external intents is that same as the sequence when the app is in the
background, which isn't the case on iOS 13+. With the exception that the
handler that starts the sequence is -applicationWillEnterForeground
instead of -applicationDidLaunchingWithOptions.

Validated the following intents handlers:
  application:openURL
  application:continueUserActivity
  application:willContinueUserActivityWithType

Wasn't able to test for: application:performActionForShortcutItem
  > But this is likely that the same startup sequence will be used as
    for the other intents.

Bug: 1205553
Change-Id: I6f1f73c329fbf1718d7664068d7b318f3c6fe175
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2872945
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/master@{#879368}
NOKEYCHECK=True
GitOrigin-RevId: 28f3b4e75c9c78ba7681c1093cac52d2235bbb32