Add bookmarks to tab grid item context menu
The Add to Bookmarks items allows for a single tab displayed on the Tab
Grid to be added to a user's bookmarks directly.

For new bookmarks, a message is displayed allowing the user to edit the
bookmark. If the URL is already bookmarked, the edit page will be
displayed immediately. (This matched the behavior of the Add to
Bookmarks action when the user is viewing a single webpage.)

Bug: 1196963
Change-Id: I306bd9b11975aefd1f6c0f10a9f88b7292a7accf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2859673
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Auto-Submit: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: edchin <edchin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#879389}
NOKEYCHECK=True
GitOrigin-RevId: b89717515f10b02f72a5605aa15da561d4fb3d25