Wire in account-scoped AutofillWebDataService on iOS
Simply call the existing GetAutofillWebDataForAccount() and pass it to
PersonalDataManager::Init(). The account-scoped wallet storage will now
be available on iOS if the kAutofillEnableAccountWalletStorage feature
is enabled. Also use the occasion to expose a chrome://flags entry for
the feature, with the same name and description as the //chrome one.

Bug: 1206190
Change-Id: I8d2695350598cdbfa71541fef7f1adf0ceef78fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2874664
Reviewed-by: Marc Treib <treib@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Commit-Queue: Victor Vianna <victorvianna@google.com>
Cr-Commit-Position: refs/heads/master@{#880356}
NOKEYCHECK=True
GitOrigin-RevId: e427a722c5c59761ac1d34611c5ee182af2db6e8