Add WebStatePolicyDecider::ShouldAllowErrorPageToBeDisplayed() method
This CL adds a new method called ShouldAllowErrorPageToBeDisplayed() to
WebStatePolicyDecider. This method is used to decide whether the
response is allowed to display an error page such as an SSL interstitial
or a DNS resolution error.

Presently, an error page causes a short circuit in the navigation path
and prevents WebStatePolicyDecoder::ShouldAllowResponse() from being
called.

With this change, if ShouldAllowErrorPageToBeDisplayed() returns false,
the navigation will still fail, but
WebStatePolicyDecider::ShouldAllowResponse() can now observe that it
failed and take an action.

The main use case for this new method is the implementation of HTTP
fallbacks in the Defaulting Omnibox Navigations to HTTPS feature. This
feature defaults typed omnibox navigations to use an HTTPS scheme and
observes the navigation. If the navigation results in a net or SSL error,
the feature immediately falls back to the HTTP URL without displaying an
error page. Presently, this isn't possible on iOS because
WebStatePolicyDecoder::ShouldAllowResponse() can't respond to navigation
failures. This feature will override ShouldAllowErrorPageToBeDisplayed()
to allow failed navigations to suppress error pages.

Bug: 1195440
Change-Id: Ic7397da76805a3beb9edf3a43dc5dec45c299640
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2802754
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881098}
NOKEYCHECK=True
GitOrigin-RevId: 0679c91c7642c5ed1f2caeac33b7a45977680f8d