[ios] Makes the allocator shim resilient against vm_protect() errors.
On iOS 14.3 iPad Air 2, the region containing the default malloc zone is
not marked as VM_PROTECT_WRITE, so it is not possible to unprotect this
region and install the allocator shim. Since this code runs
unconditionally at startup, this CL gracefully handles this and other
potential errors instead of calling MACH_CHECK().

This CL also disables the heap profiler in this partially-shimmed state,
to avoid polluting the profiler data, and logs a histogram to allow us
to determine if this ever starts failing on a future OS/device.

BUG=1154656,1150599

Change-Id: I35ec903911cb648666dbde52369813b1df1f985a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2644529
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Reviewed-by: Mark Mentovai <mark@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Erik Chen <erikchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881111}
NOKEYCHECK=True
GitOrigin-RevId: d0765563c4ae47dbd61f60a7b0fcaef9def5e4cc