[iOS] Skip generating legacy TLS error for WebStates not in a WebStateList
When a WebState isn't in a WebStateList, it doesn't have tab
helpers, and in particular, doesn't have an IOSBlockingPageTabHelper.

Attempting to generate a legacy TLS error page in such a WebState
leads to crash, while trying to use the null IOSBlockingPageTabHelper.

Since such WebStates aren't being shown to the user in the content
area, there's no reason to generate the error page text.

To prevent this crash, this CL makes GetLegacyTLSErrorPageHTML
return the empty string for WebStates that don't have tab helpers.

Bug: 1207874
Change-Id: I135d8e741c395b877e10095cb0fc356e854c4d55
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2885961
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#881719}
NOKEYCHECK=True
GitOrigin-RevId: 2072ab9440de97570f8f21b3b2802c91d3a627c9