[sync] Add method type hint to Javascript API for account recovery
The integer acts as enum but is considered opaque in most layers,
plumbed from the Javascript API in the renderer process, until it
reaches the platform-dependent TrustedVaultClient in the browser
process.

This value is expected to be useful in interactions with the server,
acting as a hint that allows the server to implement optimizations.

Change-Id: Id90aec0765129be2c1dd7a68f6ea5af5f2972854
Bug: 1081650
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2879995
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Nico Weber <thakis@chromium.org>
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Mikel Astiz <mastiz@chromium.org>
Cr-Commit-Position: refs/heads/master@{#882992}
NOKEYCHECK=True
GitOrigin-RevId: ec15f2e9fd4a6165a19560b1e070dc80d349de3d