[iOS] Prevent pdf generation when a dialog is running
When a user takes a screenshot, he can have 2 possibilities on top of
the screen:
* Screen
* Full Page

If a website display an alert, we are not able to create a full page
pdf.  Currently, when the user click on full page option at the top of
the screen, nothing is displayed. To avoid that, we can send a nil to
Apple's API so the full page option is disabled.

Bug: 1136528
Change-Id: I4f1ee39ea4d1e338920d48d0c9ab740eb5d194a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2874529
Commit-Queue: Aliona Dangla <alionadangla@google.com>
Reviewed-by: Sebastien Lalancette <seblalancette@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#883496}
NOKEYCHECK=True
GitOrigin-RevId: 4982ea4e80a43fb7b836a2c1f16542866de45b1f