Remove HTTPBad Field Edit detection
We no longer need to track whether the user has edited a form field to
decide to mark non-secure pages as /!\ Not Secure. Remove all of the
code previously used to bubble such events up to Chrome for use in
calculation of the security state of the page.

Bug: 1191143
Change-Id: I0b6ee90d299fe2e552b1c11b486bb8aecc0747cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2908178
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Carlos IL <carlosil@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Eric Lawrence [MSFT] <ericlaw@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#886377}
NOKEYCHECK=True
GitOrigin-RevId: 59a2905709fc575b31b588857c4d40583617c1b0