[iOS][Default Promo] Change lookback window to 21 days
This will help to lower the chances of a repeat, irrelevant promo.

This will change the definition of a default user to some that opened a
system URL in the last 21 days instead of 7. This shift in the
definition of "no longer default user" will increase the default count,
though we don't think it is an artificial inflation.

Bug: N/A
Change-Id: I31386760fd17b681580d830e26a65c164313dcd9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2916728
Commit-Queue: Javier Flores <javierrobles@chromium.org>
Reviewed-by: Mohammad Refaat <mrefaat@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886495}
NOKEYCHECK=True
GitOrigin-RevId: f707e57fa45398c2297cb94453453eb6b176a61c