Move SnackbarCoordinator to TabGrid
With the addition of context menu actions to tab grid cells which may
initiate a snackbar message, the snackbar presentation needs to be
handled at the tab grid level. This is necessary in order for snackbar
messages to be associated with the correct window. iPads can display
multiple windows of Chrome, but the BVC may not be presented in any of
them. Using a nil window to present the snackbar message will cause it
to be presented in the newest window rather than the one associated with
the correct Browser instance.

Fixed: 1208273
Change-Id: Ibcbe1f4412056d59d96e075e11c31527f20fdd6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2918247
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#886872}
NOKEYCHECK=True
GitOrigin-RevId: aa622e9fc8000a9a48b09118f7f4fe3cc62457e3