[iOS] Remove force-opening the NTP automatically for firstRun
shouldOpenNTPTabOnActivationOfBrowser is likely to return true when
showing First Run because the web states list is likely to be empty
(the browser is opened for the first time?). Hence, the new tab page
will be opened without the need of forcing the opening of the NTP when
|firstRun| == YES.

If, in a unlikely situation, shouldOpenNTPTabOnActivationOfBrowser
returns NO, when First Run is needed, this means that there is a tab
that has to be displayed in the browser, in which case there will still
be a browser tab view behind the FRE, maybe not a NTP, but there
will be something.

This will allow simplifying the logic to introduce the FRE agent where
there will likely be 2 new init stages: one before FRE that initialize
all the regular browser startup UI, which includes opening NTP if needed,
then the FRE stage that will show the FRE if needed on top of the NTP.

This will remove the need for the FRE agent to open a NTP.

Note, that if we really want to open a NTP when
|firstRun| == YES, we could consider having the agent setting a scene
property that indicates that the app will go through the FRE
before reaching the stage to initialize the UI behind the FRE.

Bug: 1178821
Change-Id: Iae3c7e7f7bf751a845c82a16ea04ebc2ca1615fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2905915
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/master@{#886947}
NOKEYCHECK=True
GitOrigin-RevId: c5718b8241ad752696c9c3a8ab252607bbc89801