[iOS] Introduce the FirstRun agent
This is the initial CL to move the First Run logic from Scene Controller
to an agent.

For more context see the full integration prototype crrev.com/c/2898327.

Here we introduce the FirstRunAppStateAgent agent that has already the
logic to show the FRE when needed.

The code for showing the FRE UI is still in Scene Controller, which will
be moved to the agent in a follow-up CL.

Bug: 1210251
Change-Id: Id9bdcf62fc625bb706c7df3716d816e147bc9405
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2920870
Commit-Queue: Vincent Boisselle <vincb@google.com>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888291}
NOKEYCHECK=True
GitOrigin-RevId: e53f664c6c6ba474a76953a04b80c249d7b53bae