[iOS] Fixing table view error in recents related to sign-in
SectionIdentifierOtherDevices is removed and added again at each table
view reload. With the enterprise policy to block sign-in introduced with
crrev.com/c/2799673, it is possible not add back the other device
section.

To avoid metrics issues (some metrics are recorded each time the sign-in
promo view mediator is created), the best fix is to test if the section
exists when the promo view has to be refreshed.
If the section doesn't exist, the mediator can safely be removed.

Bug: 1215062
Change-Id: Iccbe4ef7eabc8cc5e87d4bbad4a62d4f38c09a42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2933960
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jérôme <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888825}
NOKEYCHECK=True
GitOrigin-RevId: b2aea160522cab9c4caeb464b9829103369cc0c1