Pass the sign-in primary account when loading the tokens.
After this change, the token of the primary sign-in not syncing account
should no longer be revoke while the tokens are loaded. This CL thus
simplifies the logic in the SigninManager to only update the
unconsented primary account once the accounts are loaded.

Bug: 1196596

Change-Id: I6020b60c92194a7a21ce7f3b23b77b055a6347b6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2930170
Reviewed-by: David Roger <droger@chromium.org>
Reviewed-by: Jérôme <jlebel@chromium.org>
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888847}
NOKEYCHECK=True
GitOrigin-RevId: f68997e78b59ed524d0a4dfbec73984895126949