[ios] Convert OnBlockingPageCommand to use base::Value API
The base::DictionaryValue class is deprecated in favor of simply
using base::Value (and the API that internally check the stored
value to be a dictionary).

Also move to IOSBlockingPageTabHelper::OnBlockingPageCommand the
code to extract the command as a SecurityInterstitialCommand (it
was duplicated in all implementation of HandleScriptCommand).

Bug: 1187061
Change-Id: I8769e3c12c8d0d55f59504e1584e9d8635709776
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2933498
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888858}
NOKEYCHECK=True
GitOrigin-RevId: 3447d0a43e47d7bd70c5230f73d968a3b3ddf6f4