[iOS][web] Create pending item for POST form submission
When a form post a submission to the same page using the POST method,
if the destination page has the same URL as the form URL, no pending
item is created as the navigation is treated as a same URL navigation.

However, as it is a POST, a new navigation entry is created in the
navigation stack. Because of that, the information linked to this
navigation (transition type, post method...) are actually lost as
the NavigationItem associated with it was destroyed.

This CL makes sure that a new pending item is created in this case.

Fixed: 1215913
Change-Id: I69f10bf0769a699787a4462a5ec9439053a7834e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2935285
Reviewed-by: Ali Juma <ajuma@chromium.org>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#888894}
NOKEYCHECK=True
GitOrigin-RevId: 70bffef5434c910ee88c70cfb9663121a606412c