[iOS FRE] Add Enterprise strings to the welcome screen
Adds logic to the Welcome Screen in the new FRE to detect whether the
browser is being managed, and if so, change the title and subtitle to
Enterprise-oriented messaging. An enterprise icon, as well as a
disclaimer that the browser is managed by the organization, will also
be added in that case.

Bug: 1189815
Change-Id: Id6134cc14b176ea7669c0883a7207144c1c71864
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2935538
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#889373}
NOKEYCHECK=True
GitOrigin-RevId: 3c2481925405e34ae1abffb1f77f33da48a83ad2