[omnibox] [rich-autocompletion] IME compatibility.
Moves `omnibox_additional_text_view_` after
`ime_inline_autocomplete_view_`. This fixes the case when there is title
AC with IME which would previously display the URL in between the input
and autocompletion. E.g., input 'あ' and suggestion title 'ああ':
  Before: あ    (URL.com)[あ]
  Now:    あ[あ]    (URL.com)

Adds `ime_prefix_autocomplete_view_` before `omnibox_view_`. This fixes
the case when there is non prefix AC with IME which would previously
hide the prefix autocompletion and calculate the suffix autocompletion
as if the user input matched the suggestion prefix. E.g., input 'あ' and
suggestion 'url.com/ああ':
  - Before: あ[rl.com/ああ]
  - Now:    [url.com/]あ[あ]

Simplifies location bar layout calculations. Previously:
  1) The entire width would be allocated for the omnibox
  2) The omnibox would be resized to its required width if there was IME
autocompletion
  3) The 2 views would be further adjusted if there was insufficient
width to completely display both
  4) Additional processing would be done to split the allocated omnibox
width between itself and `omnibox_additional_text_view_`.

This approach had issues:
  - It was difficult to scale for the new
`ime_prefix_autocomplete_view_`
  - It was unclear if steps (3) and (4) worked correctly together; i.e.
the case when there was title AC, IME AC, and a small location bar
width.
  - It incorrectly preferred IME autocompletion over the user input when
there was a small location bar. This was inconsistent with non-IME
autocompletion.

The new approach:
  1) If the additional views are not visible, allocate the entire width
for the omnibox and skip the following steps. Otherwise:
  2) Distribute the available width between the 4 views with correct
preference when there's insufficient width to display all.
  3) Place them adjacently left to right.

Step 1 in the new approach helps reduce the cases when the omnibox
scroll is dirtied due to the omnibox resizing. It's not foolproof; e.g.,
it's not possible to skip resizing the omnibox when IME or title
autocompletion are present. But it is the status quo; i.e. it doesn't
make the edge cases worse or more frequent. I'll explore a more robust
alternative in followup CL.

Removes the omnibox's border. Previously, the omnibox had a 3px border
on all sides. Removing the border doesn't effect it when there is no IME
autocompletion; importantly, the omnibox text will continue to be
correctly left-aligned with the suggestion text. However, removing this
border fixes positioning when there is IME autocompletion. This was
especially noticeable when beginning IME input would shift the input 3px
and the autocompletion 6px to the right. E.g., input 'あ' and suggestion
'あああ':
  - Before: [あ] あ [あ]
  - Now:    [あ]あ[あ]

Bug: 1062446
Change-Id: Idf6d208afd119e9e08a4762761ef853d3a173d04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2924246
Reviewed-by: Stepan Khapugin <stkhapugin@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#890828}
NOKEYCHECK=True
GitOrigin-RevId: b643eb3be44781ea8ea28ec46fc6af25264389e7