[iOS]🐻[FRE] Handle SignIn policy change in FRE
This CL makes sure that if the SignIn policy is updated while the Sign
In or Sync screen is displayed, the user is notified.

For that, the CL is modifying the behaviour of the UserPolicySignout
coordinator, allowing it to let its delegate know when it has been
dismissed.

Bug: 1189836
Change-Id: Ic7a23aa28caef7bc0733fc5396223a305b8df38f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2947009
Reviewed-by: Guillaume Jenkins <gujen@google.com>
Commit-Queue: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#890833}
NOKEYCHECK=True
GitOrigin-RevId: fcd97a7b38953da97b4e56e0c434936014de5bce