[iOS FRE] Change primary button style when being touched down
Adds styling to primary buttons in the new FRE when they're being
touched down. This is done by deriving from the UIButton class and
overriding setHighlighted:. Because this new button class is only
intended for the FRE primary buttons, it assumes the alpha for the non-
highlighted state will always be 1.

Bug: 1189815
Change-Id: I569f6bf964621b4c9d1bade032cc9bba6f66052e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2944841
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Cr-Commit-Position: refs/heads/master@{#890874}
NOKEYCHECK=True
GitOrigin-RevId: 4381eb80469f0224abbd4d3872a4e14192a628a3