Fix the minimal stage for creating the browser UI
The FRE is currently broken because the UI cannot be initialized because
the minimal stage is InitStageFinal whereas the startup sequence stops
at InitStageFirstRun when showing the FRE.

The issue wasn't caught by eg tests because it doesn't test the FRE in
the startup sequence, but after startup is done. This is an extra
incentive for adapting the egtest fixture.

Bug: tbd
Change-Id: Ibf946936125d5fda21ba5e8d64900f17b9139b2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2950687
Auto-Submit: Vincent Boisselle <vincb@google.com>
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891181}
NOKEYCHECK=True
GitOrigin-RevId: ee349a1eb84ef490435dd33f52e0de60c058aee6