[Safe Browsing] Remove usage of thread_utils.h in UrlCheckerImpl
thread_utils.h is a utility that allows //components/safe_browsing/core
to use //content's and //ios/web's threading primitives via different
implementations that are conditionally included in the GN target. We
are going to eliminate this utility as part of eliminating
//components/safe_browsing/core's incorrect dependences on
//components/safe_browsing/{content, ios}.

This CL removes usage of thread_utils.h in
//components/safe_browsing/core's UrlCheckerImpl. This
class executes entirely on the IO thread, but it also posts tasks to a
static method on the UI thread. In this CL, we do the following:

- Change DCHECKS about being on the IO thread in the class to usage
  of a ThreadChecker
- Change the posting of tasks on the IO thread to instead post the
  tasks on the current thread (which is the IO thread)
- Have the class take in a |ui_task_runner| constructor param that it
  uses to post tasks to the UI thread

We leave removal of thread_utils.h and test_task_environment.h in the
unittest for a followup, as their usage by the unittest is not entirely
trivial.

Bug: 1216341
Change-Id: I7dba476f101cae285feffe7bd881773d9b3eb050
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2961632
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892968}
NOKEYCHECK=True
GitOrigin-RevId: fd07c31f1c4ba6aca9913c5eefa828765a2b2b63