[iOS FRE] Add highlight to UMA checkbox on tap
Adds an animated highlight on the UMA checkbox of the Welcome screen of
the new FRE while the UMA button is being pressed. The animation is
similar to the one in the settings when tapping a checkbox cell, such
as choosing the default search engine.

Bug: 1214739
Change-Id: Iafe77b35992b4d6c105247535ad14210f58ea2ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2964334
Commit-Queue: Guillaume Jenkins <gujen@google.com>
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Cr-Commit-Position: refs/heads/master@{#892996}
NOKEYCHECK=True
GitOrigin-RevId: 920fb824f3fa02a8ca38cc5ea7643668da388123