Update snackbar coordinator for new incognito browser
When the last incognito tab is closed, the incognito Browser is
recreated. The snackbar coordinator needs to be recreated in this case
to ensure it is associated with the correct Browser instance.

TEST=
1. launch
2. open an incognito tab
3. load page1 in that incognito tab
4. close incognito tab
5. open new incognito tab
6. load page2 in new incognito tab.

Now Bookmark page from overflow/page actions menu or Share page using
the omnibox share icon.

Bug: 1220035, 1220023
Change-Id: I8f11c8a14461c4b2d83e68e967d9c36cf4aa870c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2966760
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Mike Dougherty <michaeldo@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893090}
NOKEYCHECK=True
GitOrigin-RevId: 89b9f14b6bc48ecd2509434443a6e17e7176e568