[ios] Resets identifyingIconButton actions in prepareForReuse.
TableViewTextEditCell isn't clearing the identifyingIconButton's action
in prepareForReuse, so when the cell gets reused, it ends up firing both
the old and new actions. This was exposed by a UITableView behavior
change in iOS 15, causing tests that use this button to fail.

BUG=1220798

Change-Id: I443c29fe4eb3ef66da20667a6b0a7e139420c49d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2967019
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893124}
NOKEYCHECK=True
GitOrigin-RevId: bdb1a6078401fa64ad4d0985d22054269c38e162