[Safe Browsing] Remove usage of thread_utils.h in DatabaseManager
thread_utils.h is a utility that allows //components/safe_browsing/core
to use //content's and //ios/web's threading primitives via different
implementations that are conditionally included in the GN target. We
are going to eliminate this utility as part of eliminating
//components/safe_browsing/core's incorrect dependences on
//components/safe_browsing/{content, ios}.

This CL removes usage of this utility from SafeBrowsingDatabaseManager
and its subclasses. As that class and its subclasses use both the UI
and IO threads, this requires passing SequencedTaskRunners in for both
threads. That in turn has a somewhat large ripple effect through
consumers of these classes. The changes are conceptually
straightforward, however.

Bug: 1216341
Change-Id: Ia6023a2ee058721d0d60381b93b3d7193a6af487
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2963954
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893468}
NOKEYCHECK=True
GitOrigin-RevId: fe3f55a8eb9a110f2a8e9c1da0c57c8a106378ef