Support multiple observers in MostVisitedSites.
This is a precursor CL of showing NTP tiles in Omnibox in Start surface
Finale (crbug.com/1164995). Currently, a new fetch of data can only be
triggered when the observer is set, and any observer has to create a new
MostVisitedSites instance to be able to observe. This isn't memory usage
wise.

This CL changes SetMostVisitedURLsObserver to AddMostVisitedURLsObserver
which allows multiple observers which uses the same |max_sites_num| to
register to the same MostVisitedSites instance. The MostVisitedSites
starts observing any sources' changes when the first observer is added.

Bug: 1218649
Change-Id: I7317f83896742f4fc66610e4e6fea6719db86e6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2953883
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Xi Han <hanxi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#893902}
NOKEYCHECK=True
GitOrigin-RevId: 17b850fcb270afac141e39232dfb6d9a95c5ff72