[iOS] Use the current active browser to get the webstate for FRE
Most of the coordinators for the FRE can use the main browser regardless
of the browser mode at startup (incognito or regular), but there is
a particular use case, where the webstate of the current active
interface is needed to display the sync error banner. This triggers a
crash when incognito is forced by policy (browser = incognito) because
there is no webstate in the main browser.

Bug: 1221253, 1220971
Change-Id: I2fc0a2645f4a6acde6e199404528d3cbae24c805
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2970423
Reviewed-by: Mark Cogan <marq@chromium.org>
Commit-Queue: Vincent Boisselle <vincb@google.com>
Cr-Commit-Position: refs/heads/master@{#894284}
NOKEYCHECK=True
GitOrigin-RevId: dff241c599f14708dfc97fa27ff558ced721491d