🐻[Pandora] Change transition to FRE
Currently we are showing the splash screen for longer to avoid
displaying the NTP before showing the FRE. It's align with the
behavior of legacy FRE, but it causes a issue that the splash
screen will flash before showing settings page.

The reason is that we should check if the FRE is on showing
but not the flag value.

Here is the screen record after fix:
https://drive.google.com/file/d/1CD6nr3tbbHU5hW0RJxqQb4expJnD1OM4/view?usp=sharing

Bug: 1214740,1222079
Change-Id: Ic2dcb3ded6288242ccb8b0199af6ffd72ed8f5e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2977511
Reviewed-by: edchin <edchin@chromium.org>
Reviewed-by: Vincent Boisselle <vincb@google.com>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894466}
NOKEYCHECK=True
GitOrigin-RevId: 9c5a9231fdb85bb2c0e3ffe15c9bce1827379633