[iOS] iOS 12 DarkModeDynamicColor cleanup for toolbar and tab animation
Now that iOS 12 is deprecated, DarkModeDynamicColor can be removed as it
is only needed on iOS 12 where dark mode isn't supported as a way to
color incognito screens.

This CL removes usage of this method for toolbars and the background
tab animation view.

Bug: 981889
Change-Id: I69506ce2392eb9e62643ede13bd51bb2d8b7d1f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2978554
Commit-Queue: Robbie Gibson <rkgibson@google.com>
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Cr-Commit-Position: refs/heads/master@{#894794}
NOKEYCHECK=True
GitOrigin-RevId: 17d9cd9760f562066cbfa156dfb7bacd807a8c03