Add Time to Read in ReadingListTableViewController
This CL replaces the offline page size with the estimated time to read
if the Reading List Messages feature is turned on. A new section
is also added that allows the user to toggle showing the Messages prompt.

A new ReadingList.Read.AgeOnFirstRead metric will be logged to keep
track of the time it takes for a new Reading List entry to be opened.

Bug: 1195978
Change-Id: I00c561306b116e926c8d64f8dfaa84a6fc044f8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2973352
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#896965}
NOKEYCHECK=True
GitOrigin-RevId: fc527187d651f2ba82405c6bb786314e90086a8d