Add Reading List Messages logic in ReadingListJavaScriptFeature
This CL presents a Reading List Messages in ReadingListJavaScriptFeature
if the heuristics are met. The IOS.PageAddedToReadingList UKM is
logged in the InfobarDelegate's Accept() method.

It also adds logic to save the estimated read time to existing Reading
List entries when the page loads.

Bug: 1195978
Change-Id: Ia253f4121f39baf2b8897a9142e0354da101535f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2983484
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Commit-Queue: Chris Lu <thegreenfrog@chromium.org>
Cr-Commit-Position: refs/heads/master@{#897807}
NOKEYCHECK=True
GitOrigin-RevId: b26a70fbd1d64d7fa817b963cc3034388a19c97d