[ios] Fixes flake in PolicyURLBlockingTestCase.
When the URLAllowlist/URLBlocklist policies are updated, the new
blocklist is generated on a background thread. This means that the
policy update is asynchronous, and if a page is loaded quickly enough
after policies are updated, the load may not use the latest
blocklist. This was causing some PolicyURLBlockingTestCase tests to
flake.

This CL adds EG2 helpers to wait for the blocklist to be updated before
continuing with the page load.

BUG=1226048

Change-Id: Iee8bb317001ad5031e95f157ba5021308dc2bc51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3002227
Commit-Queue: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Justin Cohen <justincohen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#898269}
NOKEYCHECK=True
GitOrigin-RevId: 6d614e092368fde4dfc55682f2fcc1eb63a896ae