[iOS] Hide full screen sign-in promo when fetching capabilities is slow
This CL avoids showing the full screen sign-in promo on iOS if fetching
the CanOfferExtendedSyncPromos account capability is slower than 200ms.
As the values of the account capabilities are cached by the Google
Identity library, the promo will be shown on the next cold start.

Bug: 1202731
Change-Id: I9cb96ddf572d101769c79b578179062384cefdcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3000788
Commit-Queue: Mihai Sardarescu <msarda@chromium.org>
Reviewed-by: Mark Cogan <marq@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Reviewed-by: Nohemi Fernandez <fernandex@chromium.org>
Cr-Commit-Position: refs/heads/master@{#898796}
NOKEYCHECK=True
GitOrigin-RevId: 0743fe3001517c1618976ac037a4f1414294fcdc