[iOS] Fix window creation by dragging Incognito tab.
The window activity used to signal that a tab was being dragged to a new
tab didn't contain any information to indicate that the dragged tab was
incognito, and SceneController always attempted to drop dragged tabs
into the regular browser's WebStateList.

This CL adds annotations to the Tab Move activity to indicate the mode
of the source browser.

However, there are two additional complexities with creating new windows
by dragging incognito tabs.

The first is that as part of the MoveTabToBrowser() process, the tab
is first detached from its original WebStateList. As soon as this
happens, the SceneController sees that the last tab has been removed
from an incognito browser. Since the moving tab has not yet been added
to its new browser, if this is the only open incognito tab,
SceneController will determine that all incognito tabs are closed, so
it should queue cleanup of incognito browsing state.

However, by the time this cleanup runs, the tab has been moved and is in
its new browser, so the cleanup will hit a runtime assertion when it
tries to clean up an incognito browser with a non-empty web state.

To handle this case, this CL adds a second check for empty incognito
browsers in the queued cleanup.

The second complexity is that the SceneController determines which BVC
to show before the external intents (which include Tab Move activities)
are processed. This means that the regular BVC will be shown, but there
will be no tabs for it to display, so it will default to displaying
about://blank.

To handle this case, this CL checks the URL activities when determining
the mode to launch the UI in. Since this happens before the URL
activities (external intents) are actually handled, it means duplicating a bit of
logic, which is ugly, but further cleanup would have bloated this CL
even more.

Several pieces of logic in SceneController are hoisted into methods for
better readability.

Two disabled EG tests relating to dragging Incognito tabs are also re-
enabled.

Bug: 1176180, 1176669
Change-Id: I74762d584ef7638aaad0a74b3dbbef81a4b669e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3006735
Commit-Queue: Mark Cogan <marq@chromium.org>
Reviewed-by: Olivier Robin <olivierrobin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#898936}
NOKEYCHECK=True
GitOrigin-RevId: 3b5c0b86c55362593e6b507be823589bcd067ae7